### 2025-05-24 #244 @kyleaoman
swiftgalaxy
Submitting Author: @kyleaoman
All current maintainers: @kyleaoman
Package Name: swiftgalaxy
One-Line Description of Package: Load in particles of a simulated galaxy, rotate coordinates, easy spherical/cylindrical coordinates, access integrated properties, and more.
Repository Link:  https://github.com/SWIFTSIM/swiftgalaxy
Version submitted: 2.1.1
Archive: 10.5281/zenodo.17382007
JOSS DOI: 10.21105/joss.09278
Version accepted: v2.2.1

---

SWIFTGalaxy is a module that extends [SWIFTSimIO](http://swiftsimio.readthedocs.org/) tailored to analyses of particles belonging to individual simulated galaxies. It inherits from and extends the functionality of the SWIFTDataset. It understands the content of halo catalogues (supported: [Velociraptor](https://ui.adsabs.harvard.edu/abs/2019PASA...36...21E/abstract), [Caesar](https://caesar.readthedocs.io/en/latest/), [SOAP](https://github.com/SWIFTSIM/SOAP)) and therefore which particles belong to a galaxy or other group of particles, and its integrated properties. The particles occupy a coordinate frame that is enforced to be consistent, such that particles loaded on-the-fly will match e.g. rotations and translations of particles already in memory. Intuitive masking of particle datasets is also enabled. Finally, some utilities to make working in cylindrical and spherical coordinate systems more convenient are also provided.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

  - Who is the target audience and what are scientific applications of this package?  The package is for scientists analysing [SWIFT](https://swift.strw.leidenuniv.nl/docs/index.html) simulations of galaxies. It facilitates extracting a subset of a simulation containing many galaxies that corresponds to a single galaxy and manipulating the data.

  - Are there other Python packages that accomplish the same thing? If so, how does yours differ? [pynbody](https://github.com/pynbody/pynbody) has some points in common, but swiftgalaxy is specifically designed for SWIFT simulations by (i) using [swiftsimio](https://github.com/SWIFTSIM/swiftsimio) as a backend and (ii) having features that leverage the particular data layout and metadata annotations of SWIFT outputs.

- [x] The package contains a `paper.md` matching [JOSS's requirements][JossPaperRequirements] with a high-level description in the package root or in `inst/`.  **The paper is now available in the `joss-paper` branch.**
- [x] The package is deposited in a long-term repository with the DOI: [10.5281/zenodo.15502355](https://doi.org/10.5281/zenodo.15502355)

and editors to fill this out. **completed on a previous submission, I think that I don't need to complete it again?**
