### 2025-05-24 c79cd0f
Ignore docs, tests in codecov (#39)
