+++ codecov.yml
@@ -1,0 +1,10 @@
+coverage:
+  status:
+    project:
+      default:
+        target: 100%
+        threshold: 10%
+ignore:
+  - "tests"
+  - "docs"
+  - "examples"
