### 2025-05-24 203f7c0
Ignore directories properly
