### 2025-05-24 13135b2
Don't cd to run coverage.
