+++ .github/workflows/lint_and_test.yml
@@ -114,8 +114,7 @@ jobs:
         pip install -e swiftgalaxy/
     - name: Test with pytest including coverage report
       run: |
-        cd swiftgalaxy
-        pytest --cov=./ --cov-report=xml
+        pytest --cov=swiftgalaxy/ --cov-report=xml swiftgalaxy/tests
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v4
       env:
