### 2025-05-24 d62e0e8
Use current recommended action and cov commands.
