### 2025-05-24 616be6b
Update lint_and_test.yml
Try directory parameter to codecov action.
