+++ .github/workflows/lint_and_test.yml
@@ -114,8 +114,11 @@ jobs:
         pip install -e swiftgalaxy/
     - name: Test with pytest including coverage report
       run: |
-        pytest --cov=swiftgalaxy/ --cov-branch --cov-report=xml swiftgalaxy/tests
+        cd swiftgalaxy
+        pytest --cov --cov-branch --cov-report=xml
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v5
+      with:
+        directory: swiftgalaxy
       env:
         CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
