### 2025-05-24 7dadd9c
Set ignores relative to codecov working directory.
