+++ codecov.yml
@@ -5,6 +5,6 @@ coverage:
         target: 100%
         threshold: 10%
 ignore:
-  - "swiftgalaxy/tests"
-  - "swiftgalaxy/docs"
-  - "swiftgalaxy/examples"
+  - "tests"
+  - "docs"
+  - "examples"
