### 2025-05-24 a0d5e8f
Also set codecov working directory
