+++ .github/workflows/lint_and_test.yml
@@ -120,5 +120,6 @@ jobs:
       uses: codecov/codecov-action@v5
       with:
         directory: swiftgalaxy
+        working-directory: swiftgalaxy
       env:
         CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
