### 2025-05-24 e9e0b26
Add SOAP dependency repo to ignore list.
