### 2025-05-24 08f39f7
Adjust root directory for source files.
Because we have SOAP and swiftgalaxy repos side by side in CI we need to tell codecov to consider e.g. swiftgalaxy/swiftgalaxy/reader.py as swiftgalaxy/reader.py
