### 2025-05-24 d776ff2
Try to always work from root directory.
