+++ codecov.yml
@@ -5,9 +5,9 @@ coverage:
         target: 100%
         threshold: 10%
 ignore:
-  - "^swiftgalaxy/tests"
-  - "^swiftgalaxy/docs"
-  - "^swiftgalaxy/examples"
+  - "^tests"
+  - "^docs"
+  - "^examples"
   - "^SOAP"
 fixes:
   - "^swiftgalaxy/::"
