+++ tests/toysnap.py
@@ -43,8 +43,10 @@ n_dm_2 = 5000
 n_dm_b = n_dm_all - n_dm_1 - n_dm_2
 n_s_1 = 5000
 n_s_2 = 5000
+n_s_all = n_s_1 + n_s_2
 n_bh_1 = 1
 n_bh_2 = 1
+n_bh_all = n_bh_1 + n_bh_2
 m_g = 1e3 * u.msun
 T_g = 1e4 * u.K
 m_dm = 1e4 * u.msun
