+++ tests/test_coordinate_transformations.py
@@ -363,6 +363,18 @@ class TestManualCoordinateTransformations:
         xyz = getattr(getattr(sg, particle_name), f"{coordinate_name}")
         assert_allclose_units(xyz_before + translation, xyz, rtol=1.0e-4, atol=abstol_c)
 
+    def test_translate_warn_comoving_missing(self, sg):
+        """
+        If the translation does not have comoving information issue a warning.
+        """
+        translation = u.unyt_array(
+            [1, 1, 1],
+            units=u.Mpc,
+        )
+        msg = "Translation assumed to be in comoving"
+        with pytest.warns(RuntimeWarning, match=msg):
+            sg.translate(translation)
+
     @pytest.mark.parametrize("velocity_name", ("velocities", "extra_velocities"))
     @pytest.mark.parametrize("particle_name", present_particle_types.values())
     @pytest.mark.parametrize("before_load", (True, False))
@@ -424,6 +436,64 @@ class TestManualCoordinateTransformations:
         sg.translate(-2 * sg.metadata.boxsize)  # -2x box size
         xyz = getattr(getattr(sg, particle_name), f"{coordinate_name}")
         assert_allclose_units(xyz_before, xyz, rtol=1.0e-4, atol=abstol_c)
+
+    @pytest.mark.parametrize("coordinate_name", ("coordinates", "extra_coordinates"))
+    @pytest.mark.parametrize("particle_name", present_particle_types.values())
+    @pytest.mark.parametrize("before_load", (True, False))
+    def test_transform(self, sg, particle_name, coordinate_name, before_load):
+        """
+        Check that affine transformation works.
+        """
+        xyz_before = getattr(getattr(sg, particle_name), f"{coordinate_name}")
+        if before_load:
+            setattr(
+                getattr(sg, particle_name)._particle_dataset,
+                f"_{coordinate_name}",
+                None,
+            )
+        translation = cosmo_array(
+            [1, 1, 1],
+            units=u.Mpc,
+            comoving=True,
+            cosmo_factor=cosmo_factor(a**1, scale_factor=1.0),
+        )
+        transform = np.eye(4)
+        transform[:3, :3] = rot
+        transform[3, :3] = translation.to_comoving_value(u.Mpc)
+        sg._transform(transform)
+        xyz = getattr(getattr(sg, particle_name), f"{coordinate_name}")
+        assert_allclose_units(
+            xyz_before.dot(rot) + translation, xyz, rtol=1.0e-4, atol=abstol_c
+        )
+
+    @pytest.mark.parametrize("coordinate_name", ("velocities", "extra_velocities"))
+    @pytest.mark.parametrize("particle_name", present_particle_types.values())
+    @pytest.mark.parametrize("before_load", (True, False))
+    def test_transform_velocity(self, sg, particle_name, coordinate_name, before_load):
+        """
+        Check that affine transformation works.
+        """
+        xyz_before = getattr(getattr(sg, particle_name), f"{coordinate_name}")
+        if before_load:
+            setattr(
+                getattr(sg, particle_name)._particle_dataset,
+                f"_{coordinate_name}",
+                None,
+            )
+        translation = cosmo_array(
+            [100, 100, 100],
+            units=u.km / u.s,
+            comoving=True,
+            cosmo_factor=cosmo_factor(a**0, scale_factor=1.0),
+        )
+        transform = np.eye(4)
+        transform[:3, :3] = rot
+        transform[3, :3] = translation.to_comoving_value(u.km / u.s)
+        sg._transform(transform, boost=True)
+        xyz = getattr(getattr(sg, particle_name), f"{coordinate_name}")
[…]
