### 2025-05-26 #244 @coatless
---
---

Given the astronomy nature of this package, would you like to also pursue Astropy as well?

Please address the following:

### Package Dependency

Please emphasize the requirement for NumPy >= 2.0

https://github.com/SWIFTSIM/swiftgalaxy/blob/c79cd0f190f052f021313000c56c432f3d4740a5/pyproject.toml#L20

Please add installation instructions:

https://www.pyopensci.org/python-package-guide/documentation/repository-files/readme-file-best-practices.html#installation-instructions

Please show a **brief example** of the package in action:

https://www.pyopensci.org/python-package-guide/documentation/repository-files/readme-file-best-practices.html#brief-demonstration-of-how-to-use-the-package

Please add a community section:

https://www.pyopensci.org/python-package-guide/documentation/repository-files/readme-file-best-practices.html#a-community-section-with-links-to-contributing-guide-code-of-conduct

Please add a section that links to the gallery/alternative tutorials.

https://www.pyopensci.org/python-package-guide/documentation/repository-files/readme-file-best-practices.html#descriptive-links-to-package-documentation-short-tutorials

Please update the repository information to add the readthedocs webpage.

https://swiftgalaxy.readthedocs.io/en/latest/

Should be here:

https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/classifying-your-repository-with-topics#adding-topics-to-your-repository

(Sorry, there's only a "topics" help page for showing this... But, it's in the same location!)

### Package website

Please try to organize the landing page with:

https://www.pyopensci.org/python-package-guide/documentation/write-user-documentation/get-started.html#four-elements-of-a-good-open-source-documentation-landing-page

Example: https://www.fatiando.org/verde/latest/

Please consider placing tutorials under a clear website section, c.f.

https://www.pyopensci.org/python-package-guide/documentation/write-user-documentation/create-package-tutorials.html#sphinx-gallery

Example: https://www.fatiando.org/verde/latest/gallery/index.html#

Or, see the "User Guide" portion on the left hand side bar.

Please make available a way to obtain example data, e.g. `snap.hdf5`. This could be through a function or a simple bash script that could be run by a user to download and setup the data in an expected format.

https://swiftgalaxy.readthedocs.io/en/latest/getting_started/index.html#quick-start

### Action

Please take the necessary steps to address the README file and website requirements. We cannot send the package onto a later stage until these issues are addressed as our reviewers and your future users will need to be able to more easily navigate the documentation.
