### 2025-05-27 #244 @hamogu
This package seems to be astronomy-focussed and the source code does have astropy as a dependency. So, I wonder if it makes sense to treat this as an astropy/astronomy related package; to me it looks like that would make sense, but the authors have not set the tick-box for "astropy community".

@kyleaoman: Just checking if that's intentional or should we add it?
