### 2025-05-27 #244 @kyleaoman
@hamogu As far as I understand the astropy community packages look for use of astropy features where this is relevant, which is not really the case for this package. For example, `astropy.units` enables attaching physical units to arrays, so I think that it would be expected that a community package uses this. However, `swiftgalaxy` uses the `unyt` package from the yt project for physical units, so doesn't really fit into the astropy ecosystem in this sense. There are options to convert to astropy `Quantity` arrays if a user wants to pass output to an astropy-affiliated package (this is part of where the dependency comes in), but the connection feels weak so I left the box un-ticked. It's a similar story for other aspects such as coordinate arrays.
