### 2025-05-27 40b3658
Fine-tune codecov config and increase coverage (#40)
* Ignore directories properly

* Don't cd to run coverage.

* Use current recommended action and cov commands.

* Update lint_and_test.yml

Try directory parameter to codecov action.

* Set ignores relative to codecov working directory.

* Also set codecov working directory

* Add SOAP dependency repo to ignore list.

* Adjust root directory for source files. 

Because we have SOAP and swiftgalaxy repos side by side in CI we need to tell codecov to consider e.g. swiftgalaxy/swiftgalaxy/reader.py as swiftgalaxy/reader.py

* Try to always work
[…]
