### 2025-05-27 #244 @kyleaoman
Just so I can keep track:
- [x] Please emphasize the requirement for NumPy >= 2.0
- [x] Please add installation instructions
- [x] Please show a brief example of the package in action
- [x] Please add a community section
- [x] Please add a section that links to the gallery/alternative tutorials
- [x] Please update the repository information to add the readthedocs webpage
- [x] Please try to organize the landing page following best practice
- [x] Please consider placing tutorials under a clear website section, c.f.
- [x] Or, see the "User Guide" portion on the left hand side bar.
- [x] Please make available a way to obtain example data, e.g. snap.hdf5. This could be through a function or a simple bash script that could be run by a user to download and setup the data in an expected format.
