### 2025-05-27 57baa3e
Update citation information, add installation instructions.
