+++ README.rst
@@ -28,7 +28,9 @@ SWIFTGalaxy
 
 .. INTRO_START_LABEL
 
-SWIFTGalaxy is a module that extends SWIFTSimIO_ tailored to analyses of particles belonging to individual simulated galaxies. It inherits from and extends the functionality of the ``SWIFTDataset``. It understands the content of halo catalogues (supported: `Velociraptor`_, `Caesar`_, `SOAP`_) and therefore which particles belong to a galaxy or other group of particles, and its integrated properties. The particles occupy a coordinate frame that is enforced to be consistent, such that particles loaded on-the-fly will match e.g. rotations and translations of particles already in memory. Intuitive masking of particle datasets is also enabled. Finally, some utilities to make working in cylindrical and spherical coordinate systems more convenient are also provided.
+SWIFTGalaxy is a module that extends SWIFTSimIO_ tailored to analyses of particles belonging to individual simulated galaxies. It inherits from and extends the functionality of the ``SWIFTDataset``. It understands the content of halo catalogues (supported: `Velociraptor`_, `Caesar`_, `SOAP`_) and therefore which particles belong to a galaxy or other group of particles, and its integrated properties. The particles occupy a coordinate frame that is enforced to be consistent, such that particles loaded on-the-fly will match e.g. rotations and translations of particles already in memory. Intuitive masking of particle datasets is also enabled. Utilities to make working in cylindrical and spherical coordinate systems more convenient are also provided. Finally, tools to iterate efficiently over multiple galaxies are also provided.
+
+Installation is as simple as `pip install swiftgalaxy`. `numpy>=2.0` is required and will be installed automatically as a dependency.
 
 .. _SWIFTSimIO: http://swiftsimio.readthedocs.org
 .. _Velociraptor: https://ui.adsabs.harvard.edu/abs/2019PASA...36...21E/abstract
@@ -49,14 +51,35 @@ Citing SWIFTGalaxy
 
 .. CITING_START_LABEL
 
-.. note::
+If your use of SWIFTGalaxy leads to a publication, please cite the `swiftgalaxy entry`_ in the ASCL_ (`indexed on ADS`_). Ideally specify the version used (`Zenodo DOI`_, git commit ID and/or version number) and link to the github repository. A PyOpenSci_ review leading to a JOSS_ submission is `in progress`_, please check back for updated citation information.
 
-   ``swiftgalaxy`` will be listed on the ASCL_ after the first publication using it. Citation details will then be given here. In the meantime, if you publish a paper using ``swiftgalaxy``, please get in touch at submission time!
+.. code-block:: bibtex
+
+    @MISC{
+        2025ascl.soft05002O,
+     	author = {{Oman}, Kyle A.},
+        title = "{swiftgalaxy}",
+        keywords = {Software},
+        howpublished = {Astrophysics Source Code Library, record ascl:2505.002},
+        year = 2025,
+        month = may,
+        eid = {ascl:2505.002},
+        pages = {ascl:2505.002},
+        archivePrefix = {ascl},
+        eprint = {2505.002},
+        adsurl = {https://ui.adsabs.harvard.edu/abs/2025ascl.soft05002O},
+        adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+    }
 
 Please also consider the `citations requested for SWIFTSimIO <citeSWIFTSimIO>`_.
 
+.. _swiftgalaxy entry: https://ascl.net/2505.002
 .. _ASCL: https://ascl.net
[…]
