+++ README.rst
@@ -83,3 +83,10 @@ Please also consider the `citations requested for SWIFTSimIO <citeSWIFTSimIO>`_.
 .. _citeSWIFTSimIO: https://swiftsimio.readthedocs.io/en/latest/index.html#citing-swiftsimio
 
 .. CITING_END_LABEL
+
+Community
+---------
+
+Code contributions are very welcome! A good place to start is the [contributing guide](https://github.com/SWIFTSIM/swiftgalaxy/blob/main/CONTRIBUTING.md) and how to set up a [development environment](https://swiftgalaxy.readthedocs.io/en/latest/getting_started/index.html#installing).
+
+SWIFTGalaxy is licensed under [GPL-3.0](https://github.com/SWIFTSIM/swiftgalaxy/tree/main?tab=GPL-3.0-1-ov-file) and community members are expected to abide by the (code of conduct)[https://github.com/SWIFTSIM/swiftgalaxy/tree/main?tab=coc-ov-file].
