+++ examples/README.rst
@@ -5,4 +5,13 @@ You can view the example notebooks online using nbviewer_.
 
 .. _nbviewer: https://nbviewer.jupyter.org/
 
-- 
+- `SWIFTGalaxy demo`_
+- `Colibre quick start`_
+
+.. _SWIFTGalaxy demo: https://nbviewer.jupyter.org/github/SWIFTSIM/swiftgalaxy/blob/main/examples/SWIFTGalaxy_demo.ipynb
+.. _Colibre quick start: https://nbviewer.jupyter.org/github/SWIFTSIM/swiftgalaxy/blob/main/examples/SWIFTGalaxy_Colibre_QuickStart.ipynb
+
+The SWIFTGalaxy demo is also available as an `interactive notebook`_ on Colab_.
+
+.. _interactive notebook: https://bit.ly/3caUN7n
+.. _CoLab: https://colab.google/
