### 2025-06-03 361156f
Move data generation code to main package and add docstrings.
