+++ swiftgalaxy/demo_data.py
@@ -1,12 +1,18 @@
+"""
+Functions and definitions to retrieve, generate and use illustrative example data.
+"""
+
 import os
 import h5py
 import numpy as np
 import unyt as u
+from typing import Optional
 from astropy.cosmology import LambdaCDM
 from astropy import units as U
-from swiftsimio.objects import cosmo_array, cosmo_factor, a
-from swiftsimio import Writer
-from swiftgalaxy import MaskCollection
+from swiftsimio.objects import cosmo_array
+import swiftsimio
+from swiftsimio import Writer, SWIFTMask
+from swiftgalaxy import MaskCollection, SWIFTGalaxy
 from swiftgalaxy.halo_catalogues import _HaloCatalogue
 from swiftsimio.units import cosmo_units
 
@@ -69,44 +75,97 @@ age = u.unyt_quantity.from_astropy(
 
 
 class ToyHF(_HaloCatalogue):
+    """
+    A minimalist halo catalogue class for demo use with
+    :class:`~swiftgalaxy.reader.SWIFTGalaxy`.
+
+    Parameters
+    ----------
+    snapfile : :obj:`str`
+        The snapshot filename. (Default: ``'toysnap.hdf5'``)
+
+    index : :obj:`int`
+        The index (position in the catalogue) of the target galaxy.
+    """
 
     _index_attr = "_index"
 
-    def __init__(self, snapfile=toysnap_filename, index=0):
+    def __init__(self, snapfile: str = toysnap_filename, index: int = 0) -> None:
         self.snapfile = snapfile
         self._index = index
         super().__init__()
         return
 
-    def _load(self):
+    def _load(self) -> None:
+        """
+        Any non-trivial i/o operations needed at initialization go here.
+        """
         return
 
     @property
-    def index(self):
+    def index(self) -> int:
+        """
+        The position in the catalogue of the target galaxy.
+
+        Returns
+        -------
+        out : int
+            The position in the catalogue of the target galaxy.
+        """
         return self._mask_index()
 
-    def _generate_spatial_mask(self, SG):
-        import swiftsimio
+    def _generate_spatial_mask(self, sg: SWIFTGalaxy) -> SWIFTMask:
+        """
+        Evaluate the spatial mask (:class:`~swiftsimio.masks.SWIFTMask`) for the target
+        galaxy.
 
+        Parameters
+        ----------
+        sg : :class:`~swiftgalaxy.reader.SWIFTGalaxy`
+            The :class:`~swiftgalaxy.reader.SWIFTGalaxy` for which the mask is being
+            evaluated.
+
+        Returns
+        -------
+        out : swiftsimio.masks.SWIFTMask
+            The spatial mask.
+        """
         if self.index == 0:
             spatial_mask = cosmo_array(
                 [[centre_1 - 0.1, centre_1 + 0.1] for ax in range(3)],
                 u.Mpc,
                 comoving=True,
-                cosmo_factor=cosmo_factor(a**1, 1.0),
+                scale_factor=1.0,
+                scale_exponent=1,
             )
         elif self.index == 1:
             spatial_mask = cosmo_array(
                 [[centre_2 - 0.1, centre_2 + 0.1] for ax in range(3)],
                 u.Mpc,
                 comoving=True,
-                cosmo_factor=cosmo_factor(a**1, 1.0),
+                scale_factor=1.0,
+                scale_exponent=1,
             )
         swift_mask = swiftsimio.mask(self.snapfile, spatial_only=True)
         swift_mask.constrain_spatial(spatial_mask)
         return swift_mask
 
-    def _generate_bound_only_mask(self, SG):
+    def _generate_bound_only_mask(self, sg: SWIFTGalaxy) -> MaskCollection:
+        """
[…]
