### 2025-06-03 b315434
Fix warnings in tests.
