### 2025-06-03 dfd702b
Implement a helper class to fetch web examples.
