+++ swiftgalaxy/demo_data.py
@@ -6,7 +6,8 @@ import os
 import h5py
 import numpy as np
 import unyt as u
-from typing import Optional
+import subprocess
+from typing import Optional, Callable, Any
 from astropy.cosmology import LambdaCDM
 from astropy import units as U
 from swiftsimio.objects import cosmo_array
@@ -15,7 +16,6 @@ from swiftsimio import Writer, SWIFTMask
 from swiftgalaxy import MaskCollection, SWIFTGalaxy
 from swiftgalaxy.halo_catalogues import _HaloCatalogue
 from swiftsimio.units import cosmo_units
-
 
 my_soap_script_path = os.path.expanduser("~/code/SOAP/")
 if os.path.exists(my_soap_script_path):
@@ -27,12 +27,13 @@ except KeyError:
 soap_script = os.path.join(soap_script_path, "make_virtual_snapshot.py")
 assert os.path.exists(soap_script)
 
-toysnap_filename = "toysnap.hdf5"
-toyvr_filebase = "toyvr"
-toysoap_filename = "toysoap.hdf5"
-toysoap_membership_filebase = "toysoap_membership"
-toysoap_virtual_snapshot_filename = "toysnap_virtual.hdf5"
-toycaesar_filename = "toycaesar.hdf5"
+demo_data_dir = "demo_data"
+toysnap_filename = os.path.join(demo_data_dir, "toysnap.hdf5")
+toyvr_filebase = os.path.join(demo_data_dir, "toyvr")
+toysoap_filename = os.path.join(demo_data_dir, "toysoap.hdf5")
+toysoap_membership_filebase = os.path.join(demo_data_dir, "toysoap_membership")
+toysoap_virtual_snapshot_filename = os.path.join(demo_data_dir, "toysnap_virtual.hdf5")
+toycaesar_filename = os.path.join(demo_data_dir, "toycaesar.hdf5")
 present_particle_types = {0: "gas", 1: "dark_matter", 4: "stars", 5: "black_holes"}
 boxsize = 10.0 * u.Mpc
 n_g_all = 32**3
@@ -74,6 +75,109 @@ age = u.unyt_quantity.from_astropy(
 )
 
 
+def ensure_demo_data_directory(func: Callable) -> Callable:
+    """
+    Decorator that makes sure demo data directory exists.
+
+    Parameters
+    ----------
+    func : :obj:`Callable`
+        Function to be decorated.
+
+    Returns
+    -------
+    out : callable
+        Decorated function.
+    """
+
+    def wrapper(*args: tuple, **kwargs: dict) -> Any:
+        """
+        Create demo data directory if it does not exist, then call wrapped function.
+
+        Parameters
+        ----------
+        *args : :obj:`tuple`
+            Arguments for wrapped function.
+
+        **kwargs : :obj:`dict`
+            Keyword arguments for wrapped function.
+        """
+        if not os.path.exists(demo_data_dir):
+            os.mkdir(demo_data_dir)
+        return func(*args, **kwargs)
+
+    return wrapper
+
+
+class _WebExamples(object):
+    """
+    Helper class to fetch example data from the web storage on demand.
+    """
+
+    webstorage_location = (
+        "https://virgodb.cosma.dur.ac.uk/swift-webstorage/IOExamples/ssio_ci_04_2025/"
+    )
+    # Each example can require one or more files. The first file is the one whose path
+    # is returned when the example is requested. The other files are additional
+    # dependencies, if relevant.
+    available_examples = {
+        "snapshot": ("EagleSingle.hdf5",),
+    }
+
+    def __init__(self) -> None:
+        return
+
+    @ensure_demo_data_directory
+    def _get_webdata_if_not_present(self, filename) -> None:
+        """
[…]
