### 2025-06-04 7b02dbf
Add tests for example data helpers and cleanup demo data namespace.
