+++ tests/test_masking.py
@@ -7,14 +7,14 @@ import numpy as np
 from unyt.testing import assert_allclose_units
 from swiftgalaxy import SWIFTGalaxy, MaskCollection
 from swiftgalaxy.demo_data import (
-    create_toysnap,
-    remove_toysnap,
-    toysnap_filename,
-    present_particle_types,
-    n_g_all,
-    n_dm_all,
-    n_s_all,
-    n_bh_all,
+    _create_toysnap,
+    _remove_toysnap,
+    _toysnap_filename,
+    _present_particle_types,
+    _n_g_all,
+    _n_dm_all,
+    _n_s_all,
+    _n_bh_all,
 )
 
 abstol_nd = 1.0e-4
@@ -22,7 +22,7 @@ reltol_nd = 1.0e-4
 
 
 class TestMaskingSWIFTGalaxy:
-    @pytest.mark.parametrize("particle_name", present_particle_types.values())
+    @pytest.mark.parametrize("particle_name", _present_particle_types.values())
     @pytest.mark.parametrize("before_load", (True, False))
     def test_reordering_slice_mask(self, sg, particle_name, before_load):
         """
@@ -36,7 +36,7 @@ class TestMaskingSWIFTGalaxy:
         ids = getattr(sg, particle_name).particle_ids
         assert_allclose_units(ids_before[mask], ids, rtol=0, atol=0)
 
-    @pytest.mark.parametrize("particle_name", present_particle_types.values())
+    @pytest.mark.parametrize("particle_name", _present_particle_types.values())
     @pytest.mark.parametrize("before_load", (True, False))
     def test_reordering_int_mask(self, sg, particle_name, before_load):
         """
@@ -55,7 +55,7 @@ class TestMaskingSWIFTGalaxy:
         ids = getattr(sg, particle_name).particle_ids
         assert_allclose_units(ids_before[mask], ids, rtol=0, atol=0)
 
-    @pytest.mark.parametrize("particle_name", present_particle_types.values())
+    @pytest.mark.parametrize("particle_name", _present_particle_types.values())
     @pytest.mark.parametrize("before_load", (True, False))
     def test_bool_mask(self, sg, particle_name, before_load):
         """
@@ -92,9 +92,9 @@ class TestMaskingSWIFTGalaxy:
         logic for applying two consecutively).
         """
         try:
-            create_toysnap()
+            _create_toysnap()
             sg = SWIFTGalaxy(
-                toysnap_filename,
+                _toysnap_filename,
                 None,  # no halo_catalogue is easiest way to get no mask
                 transforms_like_coordinates={"coordinates", "extra_coordinates"},
                 transforms_like_velocities={"velocities", "extra_velocities"},
@@ -107,10 +107,10 @@ class TestMaskingSWIFTGalaxy:
             # check that cell mask is blank for all particle types:
             assert sg._spatial_mask is None
             # check that we read all the particles:
-            assert sg.gas.masses.size == n_g_all
-            assert sg.dark_matter.masses.size == n_dm_all
-            assert sg.stars.masses.size == n_s_all
-            assert sg.black_holes.masses.size == n_bh_all
+            assert sg.gas.masses.size == _n_g_all
+            assert sg.dark_matter.masses.size == _n_dm_all
+            assert sg.stars.masses.size == _n_s_all
+            assert sg.black_holes.masses.size == _n_bh_all
             # now apply an extra mask
             sg.mask_particles(MaskCollection(gas=np.s_[:1000]))
             assert sg.gas.masses.size == 1000
@@ -118,11 +118,11 @@ class TestMaskingSWIFTGalaxy:
             sg.mask_particles(MaskCollection(gas=np.s_[:100]))
             assert sg.gas.masses.size == 100
         finally:
[…]
