+++ tests/test_derived_coordinates.py
@@ -2,7 +2,7 @@ import pytest
 import numpy as np
 import unyt as u
 from unyt.testing import assert_allclose_units
-from swiftgalaxy.demo_data import present_particle_types
+from swiftgalaxy.demo_data import _present_particle_types
 from scipy.spatial.transform import Rotation
 from swiftsimio.objects import cosmo_array, cosmo_factor, a
 
@@ -12,7 +12,7 @@ abstol_a = 1.0e-4 * u.rad
 
 
 class TestCartesianCoordinates:
-    @pytest.mark.parametrize("particle_name", present_particle_types.values())
+    @pytest.mark.parametrize("particle_name", _present_particle_types.values())
     @pytest.mark.parametrize(
         "coordinate_name, mask",
         (
@@ -40,7 +40,7 @@ class TestCartesianCoordinates:
 
 
 class TestSphericalCoordinates:
-    @pytest.mark.parametrize("particle_name", present_particle_types.values())
+    @pytest.mark.parametrize("particle_name", _present_particle_types.values())
     @pytest.mark.parametrize("alias", ("r", "radius"))
     @pytest.mark.filterwarnings(
         "ignore:invalid value encountered in true_divide"
@@ -54,7 +54,7 @@ class TestSphericalCoordinates:
         r_from_cartesian = np.sqrt(np.sum(np.power(xyz, 2), axis=1))
         assert_allclose_units(spherical_r, r_from_cartesian, rtol=1.0e-4, atol=abstol_c)
 
-    @pytest.mark.parametrize("particle_name", present_particle_types.values())
+    @pytest.mark.parametrize("particle_name", _present_particle_types.values())
     @pytest.mark.parametrize("alias", ("r", "radius"))
     @pytest.mark.filterwarnings(
         "ignore:invalid value encountered in true_divide"
@@ -89,7 +89,7 @@ class TestSphericalCoordinates:
             spherical_v_r, v_r_from_cartesian, rtol=1.0e-4, atol=abstol_v
         )
 
-    @pytest.mark.parametrize("particle_name", present_particle_types.values())
+    @pytest.mark.parametrize("particle_name", _present_particle_types.values())
     @pytest.mark.parametrize("alias", ("theta", "lat", "latitude", "pol", "polar"))
     @pytest.mark.filterwarnings(
         "ignore:invalid value encountered in true_divide"
@@ -114,7 +114,7 @@ class TestSphericalCoordinates:
             spherical_theta, theta_from_cartesian, rtol=1.0e-4, atol=abstol_a
         )
 
-    @pytest.mark.parametrize("particle_name", present_particle_types.values())
+    @pytest.mark.parametrize("particle_name", _present_particle_types.values())
     @pytest.mark.parametrize("alias", ("theta", "lat", "latitude", "pol", "polar"))
     @pytest.mark.filterwarnings(
         "ignore:invalid value encountered in true_divide"
@@ -151,7 +151,7 @@ class TestSphericalCoordinates:
             spherical_v_theta, v_theta_from_cartesian, rtol=1.0e-4, atol=abstol_v
         )
 
-    @pytest.mark.parametrize("particle_name", present_particle_types.values())
+    @pytest.mark.parametrize("particle_name", _present_particle_types.values())
     @pytest.mark.parametrize("alias", ("phi", "lon", "longitude", "az", "azimuth"))
     @pytest.mark.filterwarnings(
         "ignore:invalid value encountered in true_divide"
@@ -175,7 +175,7 @@ class TestSphericalCoordinates:
             spherical_phi, phi_from_cartesian, rtol=1.0e-4, atol=abstol_a
         )
 
-    @pytest.mark.parametrize("particle_name", present_particle_types.values())
[…]
