+++ tests/test_creation.py
@@ -5,10 +5,10 @@ fundamental has gone wrong.
 
 from swiftgalaxy import SWIFTGalaxy
 from swiftgalaxy.demo_data import (
-    create_toysnap,
-    remove_toysnap,
-    toysnap_filename,
-    n_g_1,
+    _create_toysnap,
+    _remove_toysnap,
+    _toysnap_filename,
+    _n_g_1,
 )
 
 
@@ -107,17 +107,17 @@ class TestSWIFTGalaxyCreation:
         )
 
         try:
-            create_toysnap()
-            sg = SWIFTGalaxy(toysnap_filename, ToyHF())
+            _create_toysnap()
+            sg = SWIFTGalaxy(_toysnap_filename, ToyHF())
             # confirm that we loaded a namedcolumn during initialization:
             assert (
                 sg.gas.hydrogen_ionization_fractions._internal_dataset._neutral
                 is not None
             )
             # confirm that it got masked:
-            assert sg.gas.hydrogen_ionization_fractions.neutral.size == n_g_1
+            assert sg.gas.hydrogen_ionization_fractions.neutral.size == _n_g_1
         finally:
-            remove_toysnap()
+            _remove_toysnap()
             ToyHF._generate_bound_only_mask = old_generate_bound_only_mask
 
 

[…]
