### 2025-06-04 dccde18
Remove the awkward SOAP dependency.
