+++ optional_requirements.txt
@@ -3,3 +3,4 @@ black[jupyter]
 wily
 velociraptor
 git+https://github.com/dnarayanan/caesar.git
+git+https://github.com/SWIFTSIM/soap.git@soap_runtime
