+++ MANIFEST.in
@@ -1,1 +1,0 @@
-#include <filename> for non-python-code files
