+++ .github/workflows/lint_and_test.yml
@@ -47,13 +47,6 @@ jobs:
     steps:
     - name: Checkout swiftgalaxy
       uses: actions/checkout@v4
-      with:
-        path: swiftgalaxy
-    - name: Checkout SOAP (dependency)
-      uses: actions/checkout@v4
-      with:
-        repository: SWIFTSIM/SOAP
-        path: SOAP
     - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v5
       with:
@@ -71,12 +64,11 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install flake8 pytest
-        if [ -f swiftgalaxy/requirements.txt ]; then pip install -r swiftgalaxy/requirements.txt; fi
-        if [ -f swiftgalaxy/optional_requirements.txt ]; then pip install -r swiftgalaxy/optional_requirements.txt; fi
-        pip install -e swiftgalaxy/
+        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+        if [ -f optional_requirements.txt ]; then pip install -r optional_requirements.txt; fi
+        pip install -e .
     - name: Test with pytest
       run: |
-        cd swiftgalaxy
         pytest
 
   tests-and-coverage:
@@ -85,13 +77,6 @@ jobs:
     steps:
     - name: Checkout swiftgalaxy
       uses: actions/checkout@v4
-      with:
-        path: swiftgalaxy
-    - name: Checkout SOAP (dependency)
-      uses: actions/checkout@v4
-      with:
-        repository: SWIFTSIM/SOAP
-        path: SOAP
     - name: Set up Python 3.12
       uses: actions/setup-python@v5
       with:
@@ -109,12 +94,12 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install flake8 pytest pytest-cov
-        if [ -f swiftgalaxy/requirements.txt ]; then pip install -r swiftgalaxy/requirements.txt; fi
-        if [ -f swiftgalaxy/optional_requirements.txt ]; then pip install -r swiftgalaxy/optional_requirements.txt; fi
-        pip install -e swiftgalaxy/
+        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+        if [ -f optional_requirements.txt ]; then pip install -r optional_requirements.txt; fi
+        pip install -e .
     - name: Test with pytest including coverage report
       run: |
-        pytest --cov --cov-branch --cov-report=xml swiftgalaxy/tests
+        pytest --cov --cov-branch --cov-report=xml
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v5
       env:
