### 2025-06-04 ea770e5
Replace use of soap script in tests.
