+++ tests/test_halo_catalogues.py
@@ -122,17 +122,28 @@ class TestHaloCatalogues:
         """
         hf.extra_mask = "bound_only"
         if hasattr(hf, "soap_file"):
-            from swiftgalaxy.demo_data import soap_script
             import os
+            from compression.make_virtual_snapshot import make_virtual_snapshot
+            from compression.update_vds_paths import update_virtual_snapshot_paths
 
-            os.system(
-                f"python {soap_script} "
-                f"--absolute-paths "
-                f"'{_toysnap_filename}' "
-                f"'{_toysoap_membership_filebase}."
-                "{file_nr}.hdf5' "
-                f"'{_toysoap_virtual_snapshot_filename}' "
-                "0"
+            membership_filepattern = _toysoap_membership_filebase + ".{file_nr}.hdf5"
+            make_virtual_snapshot(
+                _toysnap_filename,
+                membership_filepattern,
+                _toysoap_virtual_snapshot_filename,
+                0,  # snapshot number, not used since no pattern in filenames
+            )
+            abs_snapshot_dir = os.path.abspath(os.path.dirname(_toysnap_filename))
+            abs_membership_dir = os.path.abspath(
+                os.path.dirname(membership_filepattern.format(file_nr=0))
+            )
+            abs_output_dir = os.path.abspath(
+                os.path.dirname(_toysoap_virtual_snapshot_filename)
+            )
+            rel_snapshot_dir = os.path.relpath(abs_snapshot_dir, abs_output_dir)
+            rel_membership_dir = os.path.relpath(abs_membership_dir, abs_output_dir)
+            update_virtual_snapshot_paths(
+                _toysoap_virtual_snapshot_filename, rel_snapshot_dir, rel_membership_dir
             )
             sg = SWIFTGalaxy(_toysoap_virtual_snapshot_filename, hf)
         else:
@@ -344,17 +355,28 @@ class TestHaloCataloguesMulti:
         hf_multi.extra_mask = "bound_only"
         hf_multi._mask_multi_galaxy(0)
         if hasattr(hf_multi, "soap_file"):
-            from swiftgalaxy.demo_data import soap_script
             import os
+            from compression.make_virtual_snapshot import make_virtual_snapshot
+            from compression.update_vds_paths import update_virtual_snapshot_paths
 
-            os.system(
-                f"python {soap_script} "
-                f"--absolute-paths "
-                f"'{_toysnap_filename}' "
-                f"'{_toysoap_membership_filebase}."
-                "{file_nr}.hdf5' "
-                f"'{_toysoap_virtual_snapshot_filename}' "
-                "0"
+            membership_filepattern = _toysoap_membership_filebase + ".{file_nr}.hdf5"
+            make_virtual_snapshot(
+                _toysnap_filename,
+                membership_filepattern,
+                _toysoap_virtual_snapshot_filename,
+                0,  # snapshot number, not used since no pattern in filenames
+            )
+            abs_snapshot_dir = os.path.abspath(os.path.dirname(_toysnap_filename))
+            abs_membership_dir = os.path.abspath(
+                os.path.dirname(membership_filepattern.format(file_nr=0))
+            )
+            abs_output_dir = os.path.abspath(
+                os.path.dirname(_toysoap_virtual_snapshot_filename)
+            )
+            rel_snapshot_dir = os.path.relpath(abs_snapshot_dir, abs_output_dir)
+            rel_membership_dir = os.path.relpath(abs_membership_dir, abs_output_dir)
+            update_virtual_snapshot_paths(
+                _toysoap_virtual_snapshot_filename, rel_snapshot_dir, rel_membership_dir
             )
             sg = SWIFTGalaxy(_toysoap_virtual_snapshot_filename, hf_multi)
         else:
