### 2025-06-04 6435f14
Install MPI in CI
