### 2025-06-04 9a8afb0
Also setup MPI in coverage test workflow
