+++ .github/workflows/lint_and_test.yml
@@ -79,6 +79,8 @@ jobs:
     steps:
     - name: Checkout swiftgalaxy
       uses: actions/checkout@v4
+    - name: Setup MPI
+      uses: mpi4py/setup-mpi@v1 
     - name: Set up Python 3.12
       uses: actions/setup-python@v5
       with:
