### 2025-06-05 5b588bf
Add web hosted data for caesar, soap and tests.
