### 2025-06-06 9d7cd17
Review docs and add section on demo data.
