+++ requirements.txt
@@ -1,6 +1,4 @@
 numpy
 scipy
-h5py
 unyt
 swiftsimio>=10.0.0
-astropy
