### 2025-06-06 a0fa8c3
Add a basic usage example to main readme.
