+++ requirements.txt
@@ -1,4 +1,4 @@
-numpy
+numpy>=2.0.0
 scipy
 unyt
 swiftsimio>=10.0.0
