### 2025-06-06 f3325aa
Revamp of example data (#42)
* Move data generation code to main package and add docstrings.

* Add json for sample file creation to package listing.

* Fix warnings in tests.

* Implement a helper class to fetch web examples.

* Add tests for example data helpers and cleanup demo data namespace.

* Remove the awkward SOAP dependency.

* Replace use of soap script in tests.

* Install MPI in CI

* Also setup MPI in coverage test workflow

* Add web hosted data for caesar, soap and tests.

* Add vr web data and test.

* Review docs and add section on demo data.

* Add a basic usage example to m
[…]
