+++ swiftgalaxy/demo_data.py
@@ -1,0 +1,2985 @@
+"""
+Functions and definitions to retrieve, generate and use illustrative example data.
+"""
+
+import os
+import h5py
+import numpy as np
+import unyt as u
+import subprocess
+from typing import Optional, Callable, Any, Union
+from astropy.cosmology import LambdaCDM
+from astropy import units as U
+from swiftsimio.objects import cosmo_array
+import swiftsimio
+from swiftsimio import Writer, SWIFTMask
+from swiftgalaxy import MaskCollection, SWIFTGalaxy
+from swiftgalaxy.halo_catalogues import _HaloCatalogue
+from swiftsimio.units import cosmo_units
+
+_demo_data_dir = "demo_data"
+_toysnap_filename = os.path.join(_demo_data_dir, "toysnap.hdf5")
+_toyvr_filebase = os.path.join(_demo_data_dir, "toyvr")
+_toysoap_filename = os.path.join(_demo_data_dir, "toysoap.hdf5")
+_toysoap_membership_filebase = os.path.join(_demo_data_dir, "toysoap_membership")
+_toysoap_virtual_snapshot_filename = os.path.join(
+    _demo_data_dir, "toysnap_virtual.hdf5"
+)
+_toycaesar_filename = os.path.join(_demo_data_dir, "toycaesar.hdf5")
+_present_particle_types = {0: "gas", 1: "dark_matter", 4: "stars", 5: "black_holes"}
+_boxsize = 10.0 * u.Mpc
+_n_g_all = 32**3
+_centre_1 = 2
+_centre_2 = 8
+_vcentre_1 = 200
+_vcentre_2 = 600
+_n_g_1 = 5000
+_n_g_2 = 5000
+_n_g_b = _n_g_all - _n_g_1 - _n_g_2
+_n_dm_all = 32**3
+_n_dm_1 = 5000
+_n_dm_2 = 5000
+_n_dm_b = _n_dm_all - _n_dm_1 - _n_dm_2
+_n_s_1 = 5000
+_n_s_2 = 5000
+_n_s_all = _n_s_1 + _n_s_2
+_n_bh_1 = 1
+_n_bh_2 = 1
+_n_bh_all = _n_bh_1 + _n_bh_2
+_m_g = 1e3 * u.msun
+_T_g = 1e4 * u.K
+_m_dm = 1e4 * u.msun
+_m_s = 1e3 * u.msun
+_m_bh = 1e6 * u.msun
+
+_Om_m = 0.3
+_Om_l = 0.7
+_Om_b = 0.05
+_h0 = 0.7
+_w0 = -1.0
+_rho_c = (3 * (_h0 * 100 * u.km / u.s / u.Mpc) ** 2 / 8 / np.pi / u.G).to(
+    u.msun / u.kpc**3
+)
+_age = u.unyt_quantity.from_astropy(
+    LambdaCDM(
+        H0=_h0 * 100 * U.km / U.s / U.Mpc, Om0=_Om_m, Ode0=_Om_l, Ob0=_Om_b
+    ).lookback_time(np.inf)
+)
+
+
+def _ensure_demo_data_directory(func: Callable) -> Callable:
+    """
+    Decorator that makes sure demo data directory exists.
+
+    Parameters
+    ----------
+    func : :obj:`Callable`
+        Function to be decorated.
+
+    Returns
+    -------
+    out : callable
+        Decorated function.
+    """
+
+    def wrapper(*args: tuple, **kwargs: dict) -> Any:
+        """
+        Create demo data directory if it does not exist, then call wrapped function.
+
+        Parameters
+        ----------
+        *args : :obj:`tuple`
+            Arguments for wrapped function.
+
+        **kwargs : :obj:`dict`
[…]
