+++ tests/toysnap.py
@@ -1,2341 +1,0 @@
-import os
-import h5py
-import numpy as np
-import unyt as u
-from astropy.cosmology import LambdaCDM
-from astropy import units as U
-from swiftsimio.objects import cosmo_array, cosmo_factor, a
-from swiftsimio import Writer
-from swiftgalaxy import MaskCollection
-from swiftgalaxy.halo_catalogues import _HaloCatalogue
-from swiftsimio.units import cosmo_units
-
-
-my_soap_script_path = os.path.expanduser("~/code/SOAP/")
-if os.path.exists(my_soap_script_path):
-    soap_script_path = my_soap_script_path
-try:
-    soap_script_path = os.path.join(os.environ["GITHUB_WORKSPACE"], "SOAP")
-except KeyError:
-    pass  # not on github CI
-soap_script = os.path.join(soap_script_path, "make_virtual_snapshot.py")
-assert os.path.exists(soap_script)
-
-toysnap_filename = "toysnap.hdf5"
-toyvr_filebase = "toyvr"
-toysoap_filename = "toysoap.hdf5"
-toysoap_membership_filebase = "toysoap_membership"
-toysoap_virtual_snapshot_filename = "toysnap_virtual.hdf5"
-toycaesar_filename = "toycaesar.hdf5"
-present_particle_types = {0: "gas", 1: "dark_matter", 4: "stars", 5: "black_holes"}
-boxsize = 10.0 * u.Mpc
-n_g_all = 32**3
-centre_1 = 2
-centre_2 = 8
-vcentre_1 = 200
-vcentre_2 = 600
-n_g_1 = 5000
-n_g_2 = 5000
-n_g_b = n_g_all - n_g_1 - n_g_2
-n_dm_all = 32**3
-n_dm_1 = 5000
-n_dm_2 = 5000
-n_dm_b = n_dm_all - n_dm_1 - n_dm_2
-n_s_1 = 5000
-n_s_2 = 5000
-n_s_all = n_s_1 + n_s_2
-n_bh_1 = 1
-n_bh_2 = 1
-n_bh_all = n_bh_1 + n_bh_2
-m_g = 1e3 * u.msun
-T_g = 1e4 * u.K
-m_dm = 1e4 * u.msun
-m_s = 1e3 * u.msun
-m_bh = 1e6 * u.msun
-
-Om_m = 0.3
-Om_l = 0.7
-Om_b = 0.05
-h0 = 0.7
-w0 = -1.0
-rho_c = (3 * (h0 * 100 * u.km / u.s / u.Mpc) ** 2 / 8 / np.pi / u.G).to(
-    u.msun / u.kpc**3
-)
-age = u.unyt_quantity.from_astropy(
-    LambdaCDM(
-        H0=h0 * 100 * U.km / U.s / U.Mpc, Om0=Om_m, Ode0=Om_l, Ob0=Om_b
-    ).lookback_time(np.inf)
-)
-
-
-class ToyHF(_HaloCatalogue):
-
-    _index_attr = "_index"
-
-    def __init__(self, snapfile=toysnap_filename, index=0):
-        self.snapfile = snapfile
-        self._index = index
-        super().__init__()
-        return
-
-    def _load(self):
-        return
-
-    @property
-    def index(self):
-        return self._mask_index()
-
-    def _generate_spatial_mask(self, SG):
-        import swiftsimio
-
-        if self.index == 0:
-            spatial_mask = cosmo_array(
-                [[centre_1 - 0.1, centre_1 + 0.1] for ax in range(3)],
-                u.Mpc,
-                comoving=True,
-                cosmo_factor=cosmo_factor(a**1, 1.0),
-            )
-        elif self.index == 1:
[…]
