+++ docs/source/modules/swiftgalaxy.demo_data.rst
@@ -1,0 +1,6 @@
+swiftgalaxy.halo\_catalogues module
+===================================
+
+.. automodule:: swiftgalaxy.demo_data
+   :members:
+   :show-inheritance:
