+++ requirements.txt
@@ -1,6 +1,4 @@
-numpy
+numpy>=2.0.0
 scipy
-h5py
 unyt
 swiftsimio>=10.0.0
-astropy
