+++ mypy.ini
@@ -1,3 +1,3 @@
 [mypy]
 ignore_missing_imports = True
-modules = swiftgalaxy.reader, swiftgalaxy.halo_catalogues, swiftgalaxy.masks, swiftgalaxy.iterator
+modules = swiftgalaxy.reader, swiftgalaxy.halo_catalogues, swiftgalaxy.masks, swiftgalaxy.iterator, swiftgalaxy.demo_data
