### 2025-06-07 0d148e2
Update notebook generally and to use new demo data.
