+++ docs/source/index.rst
@@ -14,7 +14,7 @@ Documentation contents
    :maxdepth: 2
 
    getting_started/index
-   demo_data/index
+   demonstration_data/index
    coordinate_transformations/index
    additional_coordinates/index
    halo_catalogues/index
