### 2025-06-07 c59d323
Add missing narrative doc file (#43)
* Rename new doc file to something not gitignored!
