### 2025-06-07 3e86e49
Swap some centre coordinates in demo data for better use with example notebooks.
