### 2025-06-07 bc48b78
Use sorcery to point to colab notebook.
