### 2025-06-07 843cb51
Add CoLab example  (#18)
* Added demo notebook

* Add README

* Fill out the current list of examples.

* Run black.

* Move example notebook to examples directory.

* Fix headings in notebook.

* Update notebook generally and to use new demo data.

* Swap some centre coordinates in demo data for better use with example notebooks.

* Add a test of example notebook.

* Use sorcery to point to colab notebook.

* Import unyt in notebook.

---------

Co-authored-by: Alejandra Aguirre-Santaella <dc-agui1@login8b.pri.cosma.local>
Co-authored-by: Kyle Oman <koman@astro.rug.nl>
Co-authored-by: Kyle Om
[…]
