+++ tests/test_halo_catalogues.py
@@ -220,11 +220,11 @@ class TestHaloCatalogues:
         """
         Check that the _centre function returns the expected centre.
         """
-        # default is minpot == 2.001 Mpc
+        # default is minpot == 2.0 Mpc
         assert_allclose_units(
             hf.centre,
             cosmo_array(
-                [_centre_1 + 0.001, _centre_1 + 0.001, _centre_1 + 0.001],
+                [_centre_1, _centre_1, _centre_1],
                 u.Mpc,
                 comoving=True,
                 cosmo_factor=cosmo_factor(a**1, 1.0),
@@ -237,11 +237,11 @@ class TestHaloCatalogues:
         """
         Check that the velocity_centre function returns the expected velocity centre.
         """
-        # default is minpot == 201. km/s
+        # default is minpot == 200. km/s
         assert_allclose_units(
             hf.velocity_centre,
             cosmo_array(
-                [_vcentre_1 + 1.0, _vcentre_1 + 1.0, _vcentre_1 + 1.0],
+                [_vcentre_1, _vcentre_1, _vcentre_1],
                 u.km / u.s,
                 comoving=True,
                 cosmo_factor=cosmo_factor(a**0, 1.0),
@@ -484,11 +484,11 @@ class TestVelociraptor:
     @pytest.mark.parametrize(
         "centre_type, expected",
         (
-            ("", _centre_1),
-            ("minpot", _centre_1 + 0.001),
+            ("", _centre_1 + 0.001),
+            ("minpot", _centre_1),
             ("mbp", _centre_1 + 0.002),
-            ("_gas", _centre_1 + 0.003),
-            ("_stars", _centre_1 + 0.004),
+            ("_gas", _centre_1 + 0.004),
+            ("_stars", _centre_1 + 0.005),
         ),
     )
     def test_centre_types(self, vr, centre_type, expected):
@@ -511,11 +511,11 @@ class TestVelociraptor:
     @pytest.mark.parametrize(
         "centre_type, expected",
         (
-            ("", _vcentre_1),
-            ("minpot", _vcentre_1 + 1.0),
+            ("", _vcentre_1 + 1.0),
+            ("minpot", _vcentre_1),
             ("mbp", _vcentre_1 + 2.0),
-            ("_gas", _vcentre_1 + 3.0),
-            ("_stars", _vcentre_1 + 4.0),
+            ("_gas", _vcentre_1 + 4.0),
+            ("_stars", _vcentre_1 + 5.0),
         ),
     )
     def test_velocity_centre_types(self, vr, centre_type, expected):
@@ -626,7 +626,7 @@ class TestCaesar:
         pass  # Caesar has nothing to do in _load
 
     @pytest.mark.parametrize(
-        "centre_type, expected", (("", _centre_1), ("minpot", _centre_1 + 0.001))
+        "centre_type, expected", (("", _centre_1 + 0.001), ("minpot", _centre_1))
     )
     def test_centre_types(self, caesar, centre_type, expected):
         """
@@ -646,7 +646,7 @@ class TestCaesar:
         )
 
     @pytest.mark.parametrize(
-        "centre_type, expected", (("", _vcentre_1), ("minpot", _vcentre_1 + 1.0))
+        "centre_type, expected", (("", _vcentre_1 + 1.0), ("minpot", _vcentre_1))
     )
     def test_vcentre_types(self, caesar, centre_type, expected):
         """
@@ -840,14 +840,14 @@ class TestSOAP:
     @pytest.mark.parametrize(
         "centre_type, expected",
         (
[…]
