### 2025-06-08 fd003aa
Target optional dependencies correctly?
