### 2025-06-08 e34f58a
Simplify docs homepage and relocate some information.
