+++ docs/source/index.rst
@@ -1,11 +1,80 @@
 SWIFTGalaxy
 ===========
 
-In a hurry? :doc:`Quick-start <getting_started/index>`.
-
 .. include:: ../../README.rst
    :start-after: INTRO_START_LABEL
    :end-before: INTRO_END_LABEL
+
+.. grid:: 1 2 1 2
+    :margin: 5 5 0 0
+    :padding: 0 0 0 0
+    :gutter: 4
+
+    .. grid-item-card:: :octicon:`info` Getting started
+        :text-align: center
+        :class-title: sd-fs-5
+        :class-card: sd-p-3
+
+        .. button-ref:: getting_started/index
+            :color: primary
+            :outline:
+            :expand:
+
+	    installation
+
+        .. button-ref:: quick-start
+            :color: primary
+            :outline:
+            :expand:
+
+	    quick start guide
+	    
+    .. grid-item-card:: :octicon:`code-square` Contributing
+        :text-align: center
+        :class-title: sd-fs-5
+        :class-card: sd-p-3
+
+        .. button-ref:: community/index
+            :click-parent:
+            :color: primary
+            :outline:
+            :expand:
+
+            information for developers
+
+	.. button-link:: https://github.com/SWIFTSIM/swiftgalaxy
+            :click-parent:
+            :color: primary
+            :outline:
+            :expand:
+
+            :octicon:`mark-github` source code :octicon:`link-external`
+
+    .. grid-item-card:: :octicon:`file-badge` Reference documentation
+        :text-align: center
+        :class-title: sd-fs-5
+        :class-card: sd-p-3
+
+        .. button-ref:: modules/index
+            :click-parent:
+            :color: primary
+            :outline:
+            :expand:
+
+	    lists of modules and functions
+
+    .. grid-item-card:: :octicon:`bookmark` Publishing?
+        :text-align: center
+        :class-title: sd-fs-5
+        :class-card: sd-p-3
+
+        .. button-ref:: citing/index
+	    :click-parent:
+            :color: primary
+            :outline:
+            :expand:
+
+	    citation information
 
 Documentation contents
 ----------------------
@@ -22,14 +91,10 @@ Documentation contents
    masking/index
    iterating/index
    visualisation/index
+   community/index
+   citing/index
    modules/index
 
-Citing SWIFTGalaxy
-------------------
-
-.. include:: ../../README.rst
-   :start-after: CITING_START_LABEL
-   :end-before: CITING_END_LABEL
 
 Indices and tables
 ------------------
