+++ docs/requirements.txt
@@ -1,5 +1,6 @@
 sphinx
 sphinx-rtd-theme
+sphinx-design
 recommonmark
 numpy
 numba
