### 2025-06-08 4953eee
Update docs homepage (#44)
* Simplify docs homepage and relocate some information.
