### 2025-06-08 717fcd4
Tweak dependencies to allow pypi upload.
