### 2025-06-08 #244 @kyleaoman
@coatless I think that I've addressed all of your requested changes now! I edited the submitted version to `2.1.1`, just made a release including the updates.
