+++ tests/test_examples.py
@@ -1,5 +1,6 @@
 import os
 import pytest
+import numpy as np
 import unyt as u
 from swiftsimio import cosmo_array, mask
 from swiftgalaxy import SWIFTGalaxy, Velociraptor, Caesar, SOAP, Standalone
@@ -13,6 +14,8 @@ from swiftgalaxy.demo_data import (
     _toysoap_membership_filebase,
     _toysoap_virtual_snapshot_filename,
     _toycaesar_filename,
+    _centre2,
+    _vcentre2,
 )
 
 
@@ -110,9 +113,17 @@ class TestGeneratedExampleData:
         """
         Check that we can create a swiftgalaxy using the helper for a generated snapshot.
         """
-        SWIFTGalaxy(
+        sg = SWIFTGalaxy(
             generated_examples.snapshot,
-            ToyHF(index=0),
+            ToyHF(index=1),
+        )
+        assert np.allclose(
+            sg.centre.to_physical_value(u.Mpc),
+            np.ones(3) * _centre2,
+        )
+        assert np.allclose(
+            sg.velocity_centre.to_physical_value(u.Mpc),
+            np.ones(3) * _vcentre2,
         )
 
     def test_velociraptor(self):
