+++ tests/test_str.py
@@ -2,6 +2,9 @@
 Tests of string representations of objects.
 """
 
+import pytest
+from swiftgalaxy.demo_data import web_examples, generated_examples
+
 
 class TestStr:
 
@@ -26,6 +29,19 @@ class TestStr:
         )
 
     def test_sg_string(self, sg):
+        """
+        Check that the swiftgalaxy has a string representation (not the one from
+        SWIFTDataset).
+        """
         string = str(sg)
         assert "SWIFTGalaxy at" in string
         assert repr(sg) == string
+
+    @pytest.mark.parametrize("demodata", [web_examples, generated_examples])
+    def test_exampledata_string(self, demodata):
+        """
+        Check that the webexample has an informative string representation.
+        """
+        string = str(demodata)
+        for k in demodata.available_examples.keys():
+            assert k in string
