### 2025-06-09 f53a95d
Implement a few more test coverage improvements.
