### 2025-06-09 4e7963f
Finish coverage of iterator module.
