+++ tests/test_iterator.py
@@ -1,4 +1,5 @@
 import pytest
+import re
 import numpy as np
 import unyt as u
 from unyt.testing import assert_allclose_units
@@ -615,3 +616,67 @@ class TestSWIFTGalaxies:
                 preload=("gas.coordinates",),  # just keep warning quiet
                 optimize_iteration="not_implemented",
             )
+
+    def test_coordinate_frame_from_and_auto_recentre_invalid(self, toysnap):
+        """
+        Check that inheriting a coordinate frame and auto-recentering are incompatible.
+        """
+        sg = SWIFTGalaxy(_toysnap_filename, ToyHF(index=0))
+        sgs = SWIFTGalaxies(
+            _toysnap_filename,
+            ToyHF(index=[0, 1]),
+            preload=("gas.coordinates",),  # just keep warning quiet
+            auto_recentre=True,
+            coordinate_frame_from=sg,
+        )
+        with pytest.raises(
+            ValueError, match="Cannot use coordinate_frame_from with auto_recentre"
+        ):
+            for sg_i in sgs:
+                pass
+
+    def test_coordinate_frame_from_in_iteration(self, toysnap):
+        """
+        Check that we can borrow a coordinate frame when iterating.
+        """
+        sg = SWIFTGalaxy(_toysnap_filename, ToyHF(index=0))
+        translation = cosmo_array(
+            [1, 0, 0],
+            u.Mpc,
+            comoving=True,
+            scale_factor=sg.metadata.a,
+            scale_exponent=1.0,
+        )
+        sg.translate(translation)
+        sgs = SWIFTGalaxies(
+            _toysnap_filename,
+            ToyHF(index=[0, 1]),
+            preload=("gas.coordinates",),  # just keep warning quiet
+            auto_recentre=False,
+            coordinate_frame_from=sg,
+        )
+        for sg_i in sgs:
+            assert np.allclose(sg.halo_catalogue.centre - translation, sg_i.centre)
+
+    def test_internal_units_mismatch_in_coordinate_frame_from(self, toysnap):
+        """
+        Check that incompatible internal units raises.
+        """
+        sg = SWIFTGalaxy(_toysnap_filename, ToyHF(index=0))
+        sg.metadata.units.length = 1 * u.kpc
+        sgs = SWIFTGalaxies(
+            _toysnap_filename,
+            ToyHF(index=[0, 1]),
+            preload=("gas.coordinates",),  # just keep warning quiet
+            auto_recentre=False,
+            coordinate_frame_from=sg,
+        )
+        with pytest.raises(
+            ValueError,
+            match=re.escape(
+                "Internal units (length and time) of coordinate_frame_from don't "
+                "match."
+            ),
+        ):
+            for sg_i in sgs:
+                pass
