### 2025-06-09 cb841fa
Don't create virtual membership for standalone SOAP catalogue.
