### 2025-06-10 #244 @crhea93
@kyleaoman My name is Carter Rhea -- I'll be the editor guiding you through this process :) Let me start off by saying that I am very excited about this package! I've used tools like yt and pyxsim for galaxy clusters.

A bit about this process:
The first thing I'll be doing is running through some basic checks. Once complete, I'll proceed with finding 2 reviewers. I'll keep you posted!
