### 2025-06-10 #244 @kyleaoman
@crhea93 pleased to meet you! (A fellow Canadian, perhaps? I'm from Ottawa.) This is my second pyopensci submission so I have some idea of the lay of the land, although it's changed a bit in the meantime. Looking forward to getting some feedback through the reviews 🙂
