### 2025-06-13 4a7888c
Add more tests to increase coverage.
