### 2025-06-16 30d5c2c
Use cosmo_quantity for scalars.
