### 2025-06-16 87e7127
Add another test and make _mask_index a property.
