### 2025-06-17 c8c8aa0
Add a few more tests and use replacement arguments for cosmo_factor.
