+++ swiftgalaxy/halo_catalogues.py
@@ -17,7 +17,7 @@ import numpy as np
 import unyt as u
 from swiftsimio import SWIFTMask, SWIFTDataset, mask
 from swiftgalaxy.masks import MaskCollection
-from swiftsimio.objects import cosmo_array, cosmo_factor, a
+from swiftsimio.objects import cosmo_array, cosmo_quantity
 
 from typing import Any, Union, Optional, TYPE_CHECKING, List, Set, Dict
 from numpy.typing import NDArray
@@ -1231,7 +1231,8 @@ class Velociraptor(_HaloCatalogue):
                 ],
                 u.Mpc,
                 comoving=True,
-                cosmo_factor=cosmo_factor(a**1, length_factor),
+                scale_factor=length_factor,
+                scale_exponent=1,
             ).squeeze()
         else:
             return cosmo_array(
@@ -1245,7 +1246,8 @@ class Velociraptor(_HaloCatalogue):
                 ],
                 u.Mpc,
                 comoving=True,
-                cosmo_factor=cosmo_factor(a**1, length_factor),
+                scale_factor=length_factor,
+                scale_exponent=1,
             )
 
     @property
@@ -1276,7 +1278,8 @@ class Velociraptor(_HaloCatalogue):
                 ],
                 u.Mpc,
                 comoving=True,
-                cosmo_factor=cosmo_factor(a**1, length_factor),
+                scale_factor=length_factor,
+                scale_exponent=1,
             ).squeeze()
         else:
             return cosmo_array(
@@ -1286,7 +1289,8 @@ class Velociraptor(_HaloCatalogue):
                 / length_factor,
                 u.Mpc,
                 comoving=True,
-                cosmo_factor=cosmo_factor(a**1, length_factor),
+                scale_factor=length_factor,
+                scale_exponent=1,
             )
 
     def _get_preload_fields(self, sg: "SWIFTGalaxy") -> Set[str]:
@@ -1337,7 +1341,8 @@ class Velociraptor(_HaloCatalogue):
                     cosmo_array(
                         getattr(self._catalogue.positions, "{:s}c".format(c)),
                         comoving=self._catalogue.units.comoving,
-                        cosmo_factor=cosmo_factor(a**1, self._catalogue.units.a),
+                        scale_factor=self._catalogue.units.a,
+                        scale_exponent=1,
                     )
                     for c in "xyz"
                 ]
@@ -1349,7 +1354,8 @@ class Velociraptor(_HaloCatalogue):
                 [0.0, 0.0, 0.0],
                 u.Mpc,
                 comoving=False,
-                cosmo_factor=cosmo_factor(a**1, self.scale_factor),
+                scale_factor=self.scale_factor,
+                scale_exponent=1,
             )
         centre = cosmo_array(
             (
@@ -1362,7 +1368,8 @@ class Velociraptor(_HaloCatalogue):
                                 "{:s}c{:s}".format(c, self.centre_type),
                             ),
                             comoving=self._catalogue.units.comoving,
-                            cosmo_factor=cosmo_factor(a**1, self._catalogue.units.a),
+                            scale_factor=self._catalogue.units.a,
+                            scale_exponent=1,
                         )
                         for c in "xyz"
                     ]
@@ -1370,7 +1377,8 @@ class Velociraptor(_HaloCatalogue):
             ).to_value(u.Mpc),
             u.Mpc,
             comoving=False,  # velociraptor gives physical centres!
-            cosmo_factor=cosmo_factor(a**1, self.scale_factor),
+            scale_factor=self.scale_factor,
+            scale_exponent=1,
         ).to_comoving()
         if self._multi_galaxy and self._multi_galaxy_catalogue_mask is None:
             return centre
[…]
