### 2025-06-17 59b6c1e
Replace more cosmo_factor with scale_factor and scale_exponent.
