+++ tests/conftest.py
@@ -2,7 +2,7 @@ import pytest
 from pathlib import Path
 import numpy as np
 import unyt as u
-from swiftsimio.objects import cosmo_array, cosmo_factor, a
+from swiftsimio.objects import cosmo_array
 from swiftgalaxy import (
     SWIFTGalaxy,
     SWIFTGalaxies,
@@ -340,19 +340,22 @@ def sa():
             [_centre_1, _centre_1, _centre_1],
             u.Mpc,
             comoving=True,
-            cosmo_factor=cosmo_factor(a**1, 1.0),
+            scale_factor=1.0,
+            scale_exponent=1,
         ),
         velocity_centre=cosmo_array(
             [_vcentre_1, _vcentre_1, _vcentre_1],
             u.km / u.s,
             comoving=True,
-            cosmo_factor=cosmo_factor(a**0, 1.0),
+            scale_factor=1.0,
+            scale_exponent=0,
         ),
         spatial_offsets=cosmo_array(
             [[-1, 1], [-1, 1], [-1, 1]],
             u.Mpc,
             comoving=True,
-            cosmo_factor=cosmo_factor(a**1, 1.0),
+            scale_factor=1.0,
+            scale_exponent=1,
         ),
     )
 
@@ -368,7 +371,8 @@ def sa_multi():
             ],
             u.Mpc,
             comoving=True,
-            cosmo_factor=cosmo_factor(a**1, 1.0),
+            scale_factor=1.0,
+            scale_exponent=1,
         ),
         velocity_centre=cosmo_array(
             [
@@ -377,13 +381,15 @@ def sa_multi():
             ],
             u.km / u.s,
             comoving=True,
-            cosmo_factor=cosmo_factor(a**0, 1.0),
+            scale_factor=1.0,
+            scale_exponent=0,
         ),
         spatial_offsets=cosmo_array(
             [[-1, 1], [-1, 1], [-1, 1]],
             u.Mpc,
             comoving=True,
-            cosmo_factor=cosmo_factor(a**1, 1.0),
+            scale_factor=1.0,
+            scale_exponent=1,
         ),
     )
 
@@ -404,19 +410,22 @@ def sg_sa():
                 [_centre_1, _centre_1, _centre_1],
                 u.Mpc,
                 comoving=True,
-                cosmo_factor=cosmo_factor(a**1, 1.0),
+                scale_factor=1.0,
+                scale_exponent=1,
             ),
             velocity_centre=cosmo_array(
                 [_vcentre_1, _vcentre_1, _vcentre_1],
                 u.km / u.s,
                 comoving=True,
-                cosmo_factor=cosmo_factor(a**0, 1.0),
+                scale_factor=1.0,
+                scale_exponent=0,
             ),
             spatial_offsets=cosmo_array(
                 [[-1, 1], [-1, 1], [-1, 1]],
                 u.Mpc,
                 comoving=True,
-                cosmo_factor=cosmo_factor(a**1, 1.0),
+                scale_factor=1.0,
+                scale_exponent=1,
             ),
         ),
     )
@@ -437,7 +446,8 @@ def sgs_sa():
                 ],
                 u.Mpc,
                 comoving=True,
-                cosmo_factor=cosmo_factor(a**1, 1.0),
+                scale_factor=1.0,
+                scale_exponent=1,
             ),
             velocity_centre=cosmo_array(
                 [
@@ -446,13 +456,15 @@ def sgs_sa():
                 ],
                 u.km / u.s,
                 comoving=True,
-                cosmo_factor=cosmo_factor(a**0, 1.0),
+                scale_factor=1.0,
+                scale_exponent=0,
             ),
[…]
