### 2025-06-17 5f0030d
Fix one more use of cosmo_quantity.
