+++ swiftgalaxy/demo_data.py
@@ -301,8 +301,9 @@ class ToyHF(_HaloCatalogue):
 
     Parameters
     ----------
-    snapfile : :obj:`str` or :class:`~pathlib._local.Path`
-        The snapshot filename. (Default: ``"demo_data/toysnap.hdf5"``)
+    snapfile : :obj:`str` or :class:`~pathlib._local.Path`, default: \
+    ``"demo_data/toysnap.hdf5"``
+        The snapshot filename.
 
     index : :obj:`int` or :obj:`list`, default: ``0``
         The index (position in the catalogue) of the target galaxy.
@@ -538,22 +539,23 @@ def _create_toysnap(
 
     Parameters
     ----------
-    snapfile : :obj:`str` or :class:`~pathlib._local.Path`
-        Filename for snapshot file. (Default: ``"demo_data/toysnap.hdf5"``)
+    snapfile : :obj:`str` or :class:`~pathlib._local.Path`, default: \
+    ``"demo_data/toysnap.hdf5"``
+        Filename for snapshot file.
 
-    alt_coord_name : :obj:`str`
+    alt_coord_name : :obj:`str`, default: ``None``
         Intended for continuous integration testing purposes. Create additional
-        coordinate-like data arrays with this name. (Default: ``None``)
+        coordinate-like data arrays with this name.
 
-    alt_vel_name : :obj:`str`
+    alt_vel_name : :obj:`str`, default: ``None``
         Intended for continuous integration testing purposes. Create additional
-        velocity-like data arrays with this name. (Default: ``None``)
+        velocity-like data arrays with this name.
 
-    alt_id_name : :obj:`str`
+    alt_id_name : :obj:`str`, default: ``None``
         Intended for continuous integration testing purposes. Create additional
-        particle ID-like data arrays with this name. (Default: ``None``)
+        particle ID-like data arrays with this name.
 
-    withfof : :obj:`bool`
+    withfof : :obj:`bool`, default: ``False``
         If ``True``, include friends-of-friends (FOF) group identifiers for each
         particle.
     """
@@ -1069,8 +1071,9 @@ def _remove_toysnap(snapfile: Union[str, Path] = _toysnap_filename) -> None:
 
     Parameters
     ----------
-    snapfile : :obj:`str` or :class:`~pathlib._local.Path`
-        Filename for snapshot file. (Default: ``"demo_data/toysnap.hdf5"``)
+    snapfile : :obj:`str` or :class:`~pathlib._local.Path`, default: \
+    ``"demo_data/toysnap.hdf5"``
+        Filename for snapshot file.
     """
     Path(snapfile).unlink(missing_ok=True)
     return
@@ -1089,9 +1092,9 @@ def _create_toyvr(filebase: Union[str, Path] = _toyvr_filebase) -> None:
 
     Parameters
     ----------
-    filebase : :obj:`str` or :class:`~pathlib._local.Path`
+    filebase : :obj:`str` or :class:`~pathlib._local.Path`, default: ``"demo_data/toyvr"``
         The base name for catalogue files (several files ``base.properties``,
-        ``base.catalog_groups``, etc. will be created). (Default: ``"demo_data/toyvr"``)
+        ``base.catalog_groups``, etc. will be created).
     """
     if not Path(f"{_toyvr_filebase}.properties").is_file():
         with h5py.File(f"{_toyvr_filebase}.properties", "w") as f:
@@ -1472,9 +1475,9 @@ def _remove_toyvr(filebase: Union[str, Path] = _toyvr_filebase) -> None:
 
     Parameters
     ----------
-    filebase : :obj:`str` or :class:`~pathlib._local.Path`
[…]
