### 2025-06-17 333450c
Increase test coverage to 100% (#45)
