+++ swiftgalaxy/demo_data.py
@@ -2,12 +2,12 @@
 Functions and definitions to retrieve, generate and use illustrative example data.
 """
 
-import os
+from pathlib import Path
 import h5py
 import numpy as np
 import unyt as u
 import subprocess
-from typing import Optional, Callable, Any, Union
+from typing import Optional, Callable, Any, Union, Sequence
 from astropy.cosmology import LambdaCDM
 from astropy import units as U
 from swiftsimio.objects import cosmo_array
@@ -17,15 +17,13 @@ from swiftgalaxy import MaskCollection, SWIFTGalaxy
 from swiftgalaxy.halo_catalogues import _HaloCatalogue
 from swiftsimio.units import cosmo_units
 
-_demo_data_dir = "demo_data"
-_toysnap_filename = os.path.join(_demo_data_dir, "toysnap.hdf5")
-_toyvr_filebase = os.path.join(_demo_data_dir, "toyvr")
-_toysoap_filename = os.path.join(_demo_data_dir, "toysoap.hdf5")
-_toysoap_membership_filebase = os.path.join(_demo_data_dir, "toysoap_membership")
-_toysoap_virtual_snapshot_filename = os.path.join(
-    _demo_data_dir, "toysnap_virtual.hdf5"
-)
-_toycaesar_filename = os.path.join(_demo_data_dir, "toycaesar.hdf5")
+_demo_data_dir = Path("demo_data")
+_toysnap_filename = _demo_data_dir / "toysnap.hdf5"
+_toyvr_filebase = _demo_data_dir / "toyvr"
+_toysoap_filename = _demo_data_dir / "toysoap.hdf5"
+_toysoap_membership_filebase = _demo_data_dir / "toysoap_membership"
+_toysoap_virtual_snapshot_filename = _demo_data_dir / "toysnap_virtual.hdf5"
+_toycaesar_filename = _demo_data_dir / "toycaesar.hdf5"
 _present_particle_types = {0: "gas", 1: "dark_matter", 4: "stars", 5: "black_holes"}
 _boxsize = 10.0 * u.Mpc
 _n_g_all = 32**3
@@ -94,8 +92,7 @@ def _ensure_demo_data_directory(func: Callable) -> Callable:
         **kwargs : :obj:`dict`
             Keyword arguments for wrapped function.
         """
-        if not os.path.exists(_demo_data_dir):
-            os.mkdir(_demo_data_dir)
+        Path(_demo_data_dir).mkdir(exist_ok=True)
         return func(*args, **kwargs)
 
     return wrapper
@@ -159,18 +156,20 @@ class _WebExamples(object):
         filename : :obj:`str`
             Name of the file to fetch from the web store.
         """
-        file_location = os.path.join(_demo_data_dir, filename)
-        if os.path.exists(file_location):
+        file_location = _demo_data_dir / filename
+        if file_location.exists():
             ret = 0
         else:
             ret = subprocess.call(
                 ["wget", f"{self.webstorage_location}{filename}", "-O", file_location]
             )
-        if ret != 0:
-            os.remove(file_location)  # it wrote an empty file, kill it
+        if ret != 0:  # pragma: no cover
+            Path(file_location).unlink(
+                missing_ok=True
+            )  # (if) it wrote an empty file, kill it
             raise RuntimeError(f"Unable to download file at {filename}")
 
-    def __getattr__(self, attr: str) -> str:
+    def __getattr__(self, attr: str) -> Path:
         """
         If the attribute is in the list of available examples get the needed file(s) and
         return the path to the requested example file.
@@ -185,16 +184,14 @@ class _WebExamples(object):
 
         Returns
         -------
-        out : :obj:`str`
+        out : :class:`~pathlib._local.Path`
[…]
