+++ tests/test_halo_catalogues.py
@@ -1,12 +1,20 @@
+from pathlib import Path
+import re
 import pytest
 import h5py
 import numpy as np
 import unyt as u
 from unyt.testing import assert_allclose_units
+from swiftsimio import mask
 from swiftgalaxy.demo_data import (
     _toysnap_filename,
+    _toysoap_filename,
+    _toyvr_filebase,
+    _toycaesar_filename,
     _toysoap_virtual_snapshot_filename,
     _toysoap_membership_filebase,
+    _create_toycaesar,
+    _remove_toycaesar,
     _n_g_1,
     _n_g_2,
     _n_g_b,
@@ -27,8 +35,14 @@ from swiftgalaxy.demo_data import (
     _present_particle_types,
 )
 from swiftgalaxy import SWIFTGalaxy, MaskCollection
-from swiftgalaxy.halo_catalogues import Velociraptor, Caesar, SOAP, Standalone
-from swiftsimio.objects import cosmo_array, cosmo_factor, a
+from swiftgalaxy.halo_catalogues import (
+    Velociraptor,
+    Caesar,
+    SOAP,
+    Standalone,
+    _MaskHelper,
+)
+from swiftsimio.objects import cosmo_array
 
 abstol_c = 1 * u.pc  # less than this is ~0
 abstol_v = 10 * u.m / u.s  # less than this is ~0
@@ -89,7 +103,8 @@ class TestHaloCatalogues:
             [[-5, 5], [-5, 5], [-5, 5]],
             u.Mpc,
             comoving=True,
-            cosmo_factor=cosmo_factor(a**1, 1.0),
+            scale_factor=1.0,
+            scale_exponent=1,
         )
         hf.extra_mask = None
         sg = SWIFTGalaxy(_toysnap_filename, hf)
@@ -122,26 +137,25 @@ class TestHaloCatalogues:
         """
         hf.extra_mask = "bound_only"
         if hasattr(hf, "soap_file"):
-            import os
             from compression.make_virtual_snapshot import make_virtual_snapshot
             from compression.update_vds_paths import update_virtual_snapshot_paths
 
-            membership_filepattern = _toysoap_membership_filebase + ".{file_nr}.hdf5"
+            membership_filepattern = (
+                str(_toysoap_membership_filebase) + ".{file_nr}.hdf5"
+            )
             make_virtual_snapshot(
                 _toysnap_filename,
                 membership_filepattern,
                 _toysoap_virtual_snapshot_filename,
                 0,  # snapshot number, not used since no pattern in filenames
             )
-            abs_snapshot_dir = os.path.abspath(os.path.dirname(_toysnap_filename))
-            abs_membership_dir = os.path.abspath(
-                os.path.dirname(membership_filepattern.format(file_nr=0))
-            )
-            abs_output_dir = os.path.abspath(
-                os.path.dirname(_toysoap_virtual_snapshot_filename)
-            )
-            rel_snapshot_dir = os.path.relpath(abs_snapshot_dir, abs_output_dir)
-            rel_membership_dir = os.path.relpath(abs_membership_dir, abs_output_dir)
+            abs_snapshot_dir = Path(_toysnap_filename).parent.absolute()
+            abs_membership_dir = Path(
+                membership_filepattern.format(file_nr=0)
+            ).parent.absolute()
+            abs_output_dir = Path(_toysoap_virtual_snapshot_filename).parent.absolute()
+            rel_snapshot_dir = abs_snapshot_dir.relative_to(abs_output_dir)
+            rel_membership_dir = abs_membership_dir.relative_to(abs_output_dir)
             update_virtual_snapshot_paths(
[…]
