+++ docs/source/conf.py
@@ -8,12 +8,12 @@
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
-# documentation root, use os.path.abspath to make it absolute, like shown here.
+# documentation root, use pathlib to make it absolute, like shown here.
 #
-import os
+from pathlib import Path
 import sys
 
-sys.path.insert(0, os.path.abspath("../.."))
+sys.path.insert(0, Path("../..").absolute())
 import swiftgalaxy
 
 
