### 2025-06-18 #244 @crhea93
## Editor comments
:wave: Hi @blakeaw and @steventgk! Thank you for volunteering to review
for pyOpenSci! <Add any additional banter here that you wish..>

Please get in touch with any questions or concerns! Your review is due: July 11, 2025
---

Reviewers: @blakeaw @steventgk
Due date: July 11, 2025
