### 2025-06-25 847e5c5
Document self arguments of wrapper functions, add missing type hints.
