### 2025-06-25 9c1efc8
Switch on enforcing type hints on all function declarations and clean up.
