+++ swiftgalaxy/masks.py
@@ -11,7 +11,9 @@ selection of particles of different types for use with
 """
 
 from copy import deepcopy
-from typing import Optional
+from typing import Optional, Union
+from types import EllipsisType
+from numpy.typing import ArrayLike
 
 
 class MaskCollection(object):
@@ -58,7 +60,9 @@ class MaskCollection(object):
 
     # Could use dataclasses module, but requires python 3.7+
 
-    def __init__(self, **kwargs) -> None:
+    def __init__(
+        self, **kwargs: Optional[Union[slice, EllipsisType, ArrayLike]]
+    ) -> None:
         for k, v in kwargs.items():
             setattr(self, k, v)
         return
