+++ mypy.ini
@@ -1,3 +1,4 @@
 [mypy]
 ignore_missing_imports = True
+disallow_untyped_defs = True
 modules = swiftgalaxy.reader, swiftgalaxy.halo_catalogues, swiftgalaxy.masks, swiftgalaxy.iterator, swiftgalaxy.demo_data
