### 2025-06-25 0eaa8ae
Add missing items in catalogue tab completion, and test (#48)
* Add missing items in catalogue tab completion, and test.

* Bump version for release and enfore version consistency.
