### 2025-06-26 c84a6df
Basic implementation of lazy masking complete.
