+++ swiftgalaxy/reader.py
@@ -1639,40 +1639,6 @@ class SWIFTGalaxy(SWIFTDataset):
             self.halo_catalogue is not None
         ):  # in server mode we don't have a halo_catalogue yet
             self._extra_mask = self.halo_catalogue._get_extra_mask(self)
-        if self._extra_mask is not None:
-            # need to mask any already loaded data
-            for particle_name in self.metadata.present_group_names:
-                if getattr(self._extra_mask, particle_name) is None:
-                    continue
-                particle_metadata = getattr(
-                    self.metadata, f"{particle_name}_properties"
-                )
-                for field_name in particle_metadata.field_names:
-                    if getattr(self, particle_name)._is_namedcolumns(field_name):
-                        named_column_dataset = getattr(
-                            getattr(self, particle_name), f"{field_name}"
-                        )._named_column_dataset
-                        for column in named_column_dataset.named_columns:
-                            data = getattr(named_column_dataset, f"_{column}")
-                            if data is None:
-                                continue
-                            setattr(
-                                named_column_dataset,
-                                f"_{column}",
-                                data[getattr(self._extra_mask, particle_name)],
-                            )
-                    else:
-                        data = getattr(
-                            getattr(self, particle_name)._particle_dataset,
-                            f"_{field_name}",
-                        )
-                        if data is None:
-                            continue
-                        setattr(
-                            getattr(self, particle_name)._particle_dataset,
-                            f"_{field_name}",
-                            data[getattr(self._extra_mask, particle_name)],
-                        )
         else:
             self._extra_mask = MaskCollection(
                 **{k: None for k in self.metadata.present_group_names}
