### 2025-06-26 7e017f9
Refactor caesar masks, pass linting.
