+++ swiftgalaxy/halo_catalogues.py
@@ -19,7 +19,7 @@ from swiftsimio import SWIFTMask, SWIFTDataset, mask
 from swiftgalaxy.masks import MaskCollection
 from swiftsimio.objects import cosmo_array, cosmo_quantity
 
-from typing import Any, Union, Optional, TYPE_CHECKING, List, Set, Dict
+from typing import Any, Union, Optional, TYPE_CHECKING, List, Set, Dict, Callable
 from numpy.typing import NDArray
 
 if TYPE_CHECKING:  # pragma: no cover
@@ -835,9 +835,34 @@ class SOAP(_HaloCatalogue):
             set of particles.
         """
 
-        def generate_lazy_mask(group_name):
+        def generate_lazy_mask(group_name: str) -> Callable:
+            """
+            Generate a function that evaluates a mask for bound particles of a specified
+            particle type. The generated function should have no parameters.
 
-            def lazy_mask():
+            Parameters
+            ----------
+            group_name : :obj:`str`
+                The particle type to evaluate a mask for.
+
+            Returns
+            -------
+            out : Callable
+                The generated function that evaluates a mask.
+            """
+
+            def lazy_mask() -> NDArray:
+                """
+                Evaluate a mask that selects bound particles by comparing the particle
+                group membership dataset ``group_nr_bound`` to the halo catalogue index.
+
+                This function must mask the data (``group_nr_bound``) that is has loaded.
+
+                Returns
+                -------
+                out : :class:`~numpy.ndarray`
+                    The mask that selects bound particles.
+                """
                 mask = getattr(
                     sg, group_name
                 )._particle_dataset.group_nr_bound.to_value(
@@ -853,13 +878,12 @@ class SOAP(_HaloCatalogue):
 
             return lazy_mask
 
-        masks = MaskCollection(
+        return MaskCollection(
             **{
                 f"_{group_name}": generate_lazy_mask(group_name)
                 for group_name in sg.metadata.present_group_names
             }
         )
-        return masks
 
     @property
     def centre(self) -> cosmo_array:
@@ -1172,14 +1196,41 @@ class Velociraptor(_HaloCatalogue):
         # because we need a lazy version and to bypass swiftgalaxy masking on read
         # while we construct the mask
 
-        def generate_lazy_mask(group_name):
+        def generate_lazy_mask(group_name: str) -> Callable:
+            """
+            Generate a function that evaluates a mask for bound particles of a specified
+            particle type. The generated function should have no parameters.
 
-            def lazy_mask():
+            Parameters
+            ----------
+            group_name : :obj:`str`
+                The particle type to evaluate a mask for.
+
+            Returns
+            -------
+            out : Callable
+                The generated function that evaluates a mask.
+            """
+
+            def lazy_mask() -> NDArray:
+                """
+                Evaluate a mask that selects bound particles by comparing the
+                ``particle_ids`` to the list of bound particle IDs.
+
+                This function must mask the data (``particle_ids``) that is has
+                loaded.
+
+                Returns
+                -------
+                out : :class:`~numpy.ndarray`
+                    The mask that selects bound particles.
+                """
                 particles = (
                     self._particles[self._multi_galaxy_catalogue_mask]
                     if self._multi_galaxy_catalogue_mask is not None
                     else self._particles
                 )
+                assert not isinstance(particles, list)  # placate mypy
                 scale_factor = (
                     particles.groups_instance.catalogue.units.a
                     if not particles.groups_instance.catalogue.units.comoving
@@ -1772,73 +1823,67 @@ class Caesar(_HaloCatalogue):
 
         cat = self._mask_catalogue()
         null_slice = np.s_[:0]  # mask that selects no particles
+        list_names = {
+            "gas": "glist",
[…]
