### 2025-06-27 603190a
Give responsibility for masking data already in memory to SWIFTGalaxy.
