+++ swiftgalaxy/halo_catalogues.py
@@ -856,8 +856,6 @@ class SOAP(_HaloCatalogue):
                 Evaluate a mask that selects bound particles by comparing the particle
                 group membership dataset ``group_nr_bound`` to the halo catalogue index.
 
-                This function must mask the data (``group_nr_bound``) that is has loaded.
-
                 Returns
                 -------
                 out : :class:`~numpy.ndarray`
@@ -870,10 +868,6 @@ class SOAP(_HaloCatalogue):
                 ) == self.input_halos.halo_catalogue_index.to_value(
                     u.dimensionless
                 )
-                # mask the group_nr_bound array that we loaded
-                getattr(sg, group_name)._particle_dataset._group_nr_bound = getattr(
-                    sg, group_name
-                )._particle_dataset._group_nr_bound[mask]
                 return mask
 
             return lazy_mask
@@ -1217,9 +1211,6 @@ class Velociraptor(_HaloCatalogue):
                 Evaluate a mask that selects bound particles by comparing the
                 ``particle_ids`` to the list of bound particle IDs.
 
-                This function must mask the data (``particle_ids``) that is has
-                loaded.
-
                 Returns
                 -------
                 out : :class:`~numpy.ndarray`
@@ -1245,10 +1236,6 @@ class Velociraptor(_HaloCatalogue):
                         scale_exponent=0,
                     ),
                 )
-                # mask the particle_ids that we loaded
-                getattr(sg, group_name)._particle_dataset._particle_ids = getattr(
-                    sg, group_name
-                )._particle_dataset._particle_ids[mask]
                 return mask
 
             return lazy_mask
@@ -1854,8 +1841,6 @@ class Caesar(_HaloCatalogue):
                 """
                 Evaluate a mask that selects bound particles by comparing the lists of
                 bound particle indices to the ranges read in the spatial mask.
-
-                This function must mask the data that is has loaded, but it loads nothing.
 
                 Returns
                 -------
