### 2025-06-27 df3cd7b
Avoid infinite recursion in lazy mask lookup.
