### 2025-06-27 54413ca
Catch all the cases where we need to mask already-loaded data.
