### 2025-06-27 c89a321
Fix a bug where None could be applied as a mask.
