### 2025-06-27 2278378
Move lazy masking logic to a helper class.
