+++ tests/test_iterator.py
@@ -245,8 +245,8 @@ class TestSWIFTGalaxies:
             )
             for ptype in _present_particle_types.values():
                 assert np.all(
-                    getattr(sg_from_sgs._extra_mask, ptype)
-                    == getattr(sg._extra_mask, ptype)
+                    getattr(sg_from_sgs._extra_mask, ptype).mask
+                    == getattr(sg._extra_mask, ptype).mask
                 )
             count += 1
         assert count == len(sgs.halo_catalogue.index)
