+++ swiftgalaxy/masks.py
@@ -11,12 +11,159 @@ selection of particles of different types for use with
 """
 
 from copy import deepcopy
-from typing import Optional, Union, Callable, TYPE_CHECKING
+from typing import Optional, Union, Callable
 from types import EllipsisType
 from numpy.typing import ArrayLike
 
-if TYPE_CHECKING:
-    from swiftgalaxy import SWIFTGalaxy
+
+class LazyMask(object):
+    """
+    A class to hold a function to evaluate a mask until it is needed.
+
+    This class can contain either an explicitly evaluated mask (boolean array,
+    slice, etc.) or a reference to a function that returns such a mask when
+    called. When the ``mask`` property is accessed, if the mask is already
+    evaluated it is returned, otherwise it is evaluated and returned.
+
+    The ``_evaluated`` attribute tracks whether the explicitly evaluated
+    mask is available.
+
+    Parameters
+    ----------
+    mask : slice, default: ``None``
+        An object that can be used to mask an array (could be a slice, boolean array, etc)
+
+    mask_function : Callable, default: ``None``
+        A reference to a function that returns a mask when called.
+    """
+
+    _mask_function: Optional[Callable]
+    _mask: Optional[Union[slice, EllipsisType, ArrayLike]]
+    _evaluated: bool
+
+    def __init__(
+        self,
+        mask: Optional[Union[slice, EllipsisType, ArrayLike]] = None,
+        mask_function: Optional[Callable] = None,
+    ):
+        if mask_function is None and mask is None:
+            self._mask = None
+            self._evaluated = True
+        elif mask_function is not None and mask is None:
+            self._mask_function = mask_function
+            self._evaluated = False
+            # leave self.mask unset
+        else:
+            self._mask = mask
+            self._mask_function = mask_function
+            self._evaluated = True
+        return
+
+    def _evaluate(self) -> None:
+        """
+        Forces evaluation the mask function.
+        """
+        assert self._mask_function is not None  # placate mypy
+        self._mask = self._mask_function()
+        self._evaluated = True
+
+    @property
+    def mask(self) -> Optional[Union[slice, EllipsisType, ArrayLike]]:
+        """
+        Get the explicitly evaluated mask, evaluating it if necessary.
+
+        Returns
+        -------
+        out : slice
+            The explicitly evaluated mask.
+        """
+        if not self._evaluated:
+            self._evaluate()
+        return self._mask
+
+    def __copy__(self) -> "LazyMask":
+        """
+        Make a copy of the :class:`~swiftgalaxy.masks.LazyMask` without copying
+        data (a "shallow" copy).
+
+        Returns
+        -------
+        out : :class:`~swiftgalaxy.masks.LazyMask`
+            The copy of the :class:`~swiftgalaxy.masks.LazyMask`.
+        """
+        if self._evaluated:
+            return LazyMask(mask=self._mask, mask_function=self._mask_function)
+        else:
+            return LazyMask(mask_function=self._mask_function)
+
+    def __deepcopy__(self, memo: Optional[dict] = None) -> "LazyMask":
+        """
+        Make a copy of the :class:`~swiftgalaxy.masks.LazyMask`, copying data
+        (a "deep" copy).
+
+        Parameters
+        ----------
+        memo : :obj:`dict` (optional), default: ``None``
+            For the copy operation to keep a record of already copied objects.
+
+        Returns
+        -------
+        out : :class:`~swiftgalaxy.masks.LazyMask`
+            The copy of the :class:`~swiftgalaxy.masks.LazyMask`.
+        """
+        if self._evaluated:
+            return LazyMask(
+                mask=deepcopy(self._mask), mask_function=deepcopy(self._mask_function)
+            )
+        else:
+            return LazyMask(mask_function=deepcopy(self._mask_function))
+
[…]
