### 2025-06-27 SWIFTSIM/swiftgalaxy#50 @blakeaw
pyOpenSci review - stale references to setup.py in CONTRIBUTING.md
Hi, this issue is part of this package's pyOpenSci review at https://github.com/pyOpenSci/software-submission/issues/244

**Severity**: Minor

## Description

There are a few references to the `setup.py` file in the contribution guidelines [CONTRIBUTING](https://github.com/SWIFTSIM/swiftgalaxy/blob/main/CONTRIBUTING.md) in the **Black style**, **Flake8 style enforcement**, and the **Pytest unit testing** sections. E.g.:

    To check your copy of the repository you can then run black --check --diff in the same directory as the setup.py file.

 I see you replaced `setup.py` with `pyproject.toml` in PR https://github.com/SWIFTSIM/swiftgalaxy/pull/3, so the stale references to `setup.py` should be updated accordingly.
