### 2025-06-28 1d429f8
Don't mask mask construction data when preparing server masks.
