+++ swiftgalaxy/reader.py
@@ -1638,8 +1638,6 @@ class SWIFTGalaxy(SWIFTDataset):
         if self.halo_catalogue is not None:
             # in server mode we don't have a halo_catalogue yet
             self._extra_mask = self.halo_catalogue._get_extra_mask(self)
-        if self._extra_mask is not None:
-            self._apply_extra_mask_to_loaded_data()
         else:
             self._extra_mask = MaskCollection(
                 **{k: None for k in self.metadata.present_group_names}
@@ -2256,58 +2254,6 @@ class SWIFTGalaxy(SWIFTDataset):
                     )
                     setattr(dataset, f"_{field_name}", field_data)
         return
-
-    def _apply_extra_mask_to_loaded_data(
-        self, particle_names: Optional[list[str]] = None
-    ) -> None:
-        """
-        Check for data already in memory and apply the extra_mask if any are found.
-
-        This should only be called immediately after evaluating the extra mask and a new
-        evaluation of a lazily-evaluated mask. In the latter case it should only be called
-        on the specific particle type that the mask was evaluated for by using the
-        ``particle_names`` keyword argument.
-
-        Parameters
-        ----------
-        particle_names : list, default: ``None``
-            List of particle types to check for data that need masking.
-        """
-        if particle_names is None:
-            particle_names = self.metadata.present_group_names
-        for particle_name in particle_names:
-            particle_metadata = getattr(self.metadata, f"{particle_name}_properties")
-            for field_name in particle_metadata.field_names:
-                if getattr(self, particle_name)._is_namedcolumns(field_name):
-                    named_column_dataset = getattr(
-                        getattr(self, particle_name), f"{field_name}"
-                    )._named_column_dataset
-                    for column in named_column_dataset.named_columns:
-                        data = getattr(named_column_dataset, f"_{column}")
-                        if data is None:
-                            continue
-                        mask = getattr(self._extra_mask, particle_name)
-                        if mask is None:
-                            continue
-                        setattr(
-                            named_column_dataset,
-                            f"_{column}",
-                            data[mask.mask],
-                        )
-                else:
-                    data = getattr(
-                        getattr(self, particle_name)._particle_dataset, f"_{field_name}"
-                    )
-                    if data is None:
-                        continue
-                    mask = getattr(self._extra_mask, particle_name)
-                    if mask is None:
-                        continue
-                    setattr(
-                        getattr(self, particle_name)._particle_dataset,
-                        f"_{field_name}",
-                        data[mask.mask],
-                    )
 
     def mask_particles(self, mask_collection: MaskCollection) -> None:
         """
