+++ swiftgalaxy/halo_catalogues.py
@@ -851,12 +851,19 @@ class SOAP(_HaloCatalogue):
                 The generated function that evaluates a mask.
             """
 
-            def lazy_mask() -> NDArray:
+            def lazy_mask(mask_loaded_data: bool = True) -> NDArray:
                 """
                 Evaluate a mask that selects bound particles by comparing the particle
                 group membership dataset ``group_nr_bound`` to the halo catalogue index.
 
                 This function must mask the data (``group_nr_bound``) that is has loaded.
+
+                Parameters
+                ----------
+                mask_loaded_data : :obj:`bool`, default ``True``
+                    If ``True``, data loaded while constructing the mask is masked
+                    during this function call. Set ``False`` when called from
+                    a :class:`~swiftgalaxy.iterator.SWIFTGalaxies` "server".
 
                 Returns
                 -------
@@ -870,10 +877,11 @@ class SOAP(_HaloCatalogue):
                 ) == self.input_halos.halo_catalogue_index.to_value(
                     u.dimensionless
                 )
-                # mask the group_nr_bound array that we loaded
-                getattr(sg, group_name)._particle_dataset._group_nr_bound = getattr(
-                    sg, group_name
-                )._particle_dataset._group_nr_bound[mask]
+                if mask_loaded_data:
+                    # mask the group_nr_bound array that we loaded
+                    getattr(sg, group_name)._particle_dataset._group_nr_bound = getattr(
+                        sg, group_name
+                    )._particle_dataset._group_nr_bound[mask]
                 return mask
 
             return LazyMask(mask_function=lazy_mask)
@@ -1212,13 +1220,20 @@ class Velociraptor(_HaloCatalogue):
                 The generated function that evaluates a mask.
             """
 
-            def lazy_mask() -> NDArray:
+            def lazy_mask(mask_loaded_data: bool = True) -> NDArray:
                 """
                 Evaluate a mask that selects bound particles by comparing the
                 ``particle_ids`` to the list of bound particle IDs.
 
                 This function must mask the data (``particle_ids``) that is has
                 loaded.
+
+                Parameters
+                ----------
+                mask_loaded_data : :obj:`bool`, default ``True``
+                    If ``True``, data loaded while constructing the mask is masked
+                    during this function call. Set ``False`` when called from
+                    a :class:`~swiftgalaxy.iterator.SWIFTGalaxies` "server".
 
                 Returns
                 -------
@@ -1245,10 +1260,11 @@ class Velociraptor(_HaloCatalogue):
                         scale_exponent=0,
                     ),
                 )
-                # mask the particle_ids that we loaded
-                getattr(sg, group_name)._particle_dataset._particle_ids = getattr(
-                    sg, group_name
-                )._particle_dataset._particle_ids[mask]
+                if mask_loaded_data:
+                    # mask the particle_ids that we loaded
+                    getattr(sg, group_name)._particle_dataset._particle_ids = getattr(
+                        sg, group_name
+                    )._particle_dataset._particle_ids[mask]
                 return mask
 
             return LazyMask(mask_function=lazy_mask)
@@ -1850,12 +1866,19 @@ class Caesar(_HaloCatalogue):
                 The generated function that evaluates a mask.
             """
 
-            def lazy_mask() -> Union[NDArray, slice]:
+            def lazy_mask(mask_loaded_data: bool = True) -> Union[NDArray, slice]:
                 """
                 Evaluate a mask that selects bound particles by comparing the lists of
                 bound particle indices to the ranges read in the spatial mask.
 
                 This function must mask the data that is has loaded, but it loads nothing.
+
+                Parameters
+                ----------
+                mask_loaded_data : :obj:`bool`, default ``True``
[…]
