### 2025-07-01 SWIFTSIM/swiftgalaxy#51 @blakeaw
pyOpenSci review - target audience in package description in README
Hi, this issue is part of this package's pyOpenSci review at https://github.com/pyOpenSci/software-submission/issues/244

**Severity**: Minor

## Description

I think it would be a good idea to add an extra note to the description to more clearly indicate the intended audience, particularly for students and folks like me who are from different domains. Pulling a little from the first sentence of your [CODE_OF_CONDUCT](https://github.com/SWIFTSIM/swiftgalaxy/blob/main/CODE_OF_CONDUCT.md), maybe something as simple as updating the first sentence in the README description to:

    SWIFTGalaxy is an open source Astronomy module that extends SWIFTSimIO...
