### 2025-07-06 dc99db3
Replace setup with pyproject in contributing
