### 2025-07-10 SWIFTSIM/swiftgalaxy#53 @blakeaw
pyOpenSci review - caesar dependency requires Cython
Hi, this issue is part of this package's pyOpenSci review at https://github.com/pyOpenSci/software-submission/issues/244

Severity: Moderate

## Description

I got an error when trying to install `caesar`: `pip install git+https://github.com/dnarayanan/caesar.git`

After installing `cython`, the `caesar` installation worked fine. I'm not sure if others have experienced this, but it may be worth adding a note to the `swiftgalaxy` installation instructions that [Cython](https://cython.org/) is needed for `caesar` when using `pip` with the `git+` link.
