### 2025-07-10 SWIFTSIM/swiftgalaxy#53 @blakeaw
However, the `caesar` installation did give a DEPRECATION warning:

>   DEPRECATION: Building 'caesar' using the legacy setup.py bdist_wheel mechanism, which will be removed in a future version. pip 25.3 will enforce this behaviour change. A possible replacement is to use the standardized build interface by setting the `--use-pep517` option, (possibly combined with `--no-build-isolation`), or adding a `pyproject.toml` file to the source tree of 'caesar'. Discussion can be found at https://github.com/pypa/pip/issues/6334

I guess this could be an issue down the line for this dependency. It might be worth pinging the package maintainers to see if they will update it.
