### 2025-07-10 SWIFTSIM/swiftgalaxy#53 @blakeaw
I also got a similar error for `caesar` when installing optional dependencies for the dev environment: `pip install -r swiftgalaxy/optional_requirements.txt`
