### 2025-07-10 SWIFTSIM/swiftgalaxy#54 @blakeaw
pyOpenSci review - numpydoc lint command for Windows powershell
Hi, this issue is part of this package's pyOpenSci review at https://github.com/pyOpenSci/software-submission/issues/244

**Severity:** Minor

### Description

The following command to manually run numpydoc linting doesn't work in Windows PowerShell:
`python -m numpydoc lint swiftgalaxy**/*.py`

I only noticed because I'm working in Powershell, but this seems to work:
```sh
 Get-ChildItem -Recurse -Filter *.py -Path swiftgalaxy | ForEach-Object {python -m numpydoc lint $_.FullName}
```

I'm not sure if it's necessary to add this to the documentation or not. Still, it may be worth considering how things in the development environment translate to other OSs.
