### 2025-07-11 SWIFTSIM/swiftgalaxy#55 @blakeaw
There were also a few `AttributeError`s:

> ERROR tests/test_halo_catalogues.py::TestSOAPWithSWIFTGalaxy::test_catalogue_exposed - AttributeError: 'DarkMatterDataset' object has no attribute 'group_nr_bound'
> ERROR tests/test_halo_catalogues.py::TestSOAPWithSWIFTGalaxy::test_masks_compatible[gas] - AttributeError: 'DarkMatterDataset' object has no attribute 'group_nr_bound'
> ERROR tests/test_halo_catalogues.py::TestSOAPWithSWIFTGalaxy::test_masks_compatible[dark_matter] - AttributeError: 'DarkMatterDataset' object has no attribute 'group_nr_bound'
> ERROR tests/test_halo_catalogues.py::TestSOAPWithSWIFTGalaxy::test_masks_compatible[stars] - AttributeError: 'DarkMatterDataset' object has no attribute 'group_nr_bound'
> ERROR tests/test_halo_catalogues.py::TestSOAPWithSWIFTGalaxy::test_masks_compatible[black_holes] - AttributeError: 'DarkMatterDataset' object has no attribute 'group_nr_bound'
