### 2025-07-11 SWIFTSIM/swiftgalaxy#55 @blakeaw
And a few `exceptiongroup.ExceptionGroup` errors related to function teardown:

> ERROR tests/test_visualisation.py::TestRecenteredVisualisation::test_recentered_projection[False-False] - exceptiongroup.ExceptionGroup: errors while tearing down <Function test_recentered_projection[False-False]> (2 sub-exceptions)
> ERROR tests/test_visualisation.py::TestRecenteredVisualisation::test_recentered_projection[False-True] - exceptiongroup.ExceptionGroup: errors while tearing down <Function test_recentered_projection[False-True]> (2 sub-exceptions)
> ERROR tests/test_visualisation.py::TestRecenteredVisualisation::test_recentered_projection[True-False] - exceptiongroup.ExceptionGroup: errors while tearing down <Function test_recentered_projection[True-False]> (2 sub-exceptions)
> ERROR tests/test_visualisation.py::TestRecenteredVisualisation::test_recentered_projection[True-True] - exceptiongroup.ExceptionGroup: errors while tearing down <Function test_recentered_projection[True-True]> (2 sub-exceptions)
> ERROR tests/test_visualisation.py::TestRecenteredVisualisation::test_recentered_slice[False] - exceptiongroup.ExceptionGroup: errors while tearing down <Function test_recentered_slice[False]> (2 sub-exceptions)
> ERROR tests/test_visualisation.py::TestRecenteredVisualisation::test_recentered_slice[True] - exceptiongroup.ExceptionGroup: errors while tearing down <Function test_recentered_slice[True]> (2 sub-exceptions)
> ERROR tests/test_visualisation.py::TestRecenteredVisualisation::test_recentered_volume_render[False] - exceptiongroup.ExceptionGroup: errors while tearing down <Function test_recentered_volume_render[False]> (2 sub-exceptions)
> ERROR tests/test_visualisation.py::TestRecenteredVisualisation::test_recentered_volume_render[True] - exceptiongroup.ExceptionGroup: errors while tearing down <Function test_recentered_volume_render[True]> (2 sub-exceptions)
