### 2025-07-11 SWIFTSIM/swiftgalaxy#55 @blakeaw
And some `FileNotFoundError`s:

> FAILED tests/test_examples.py::TestWebExampleData::test_snapshot - FileNotFoundError: [WinError 2] The system cannot find the file specified
> FAILED tests/test_examples.py::TestWebExampleData::test_caesar[halo] - FileNotFoundError: [WinError 2] The system cannot find the file specified
> FAILED tests/test_examples.py::TestWebExampleData::test_caesar[galaxy] - FileNotFoundError: [WinError 2] The system cannot find the file specified
> FAILED tests/test_examples.py::TestWebExampleData::test_soap - FileNotFoundError: [WinError 2] The system cannot find the file specified
> FAILED tests/test_examples.py::TestWebExampleData::test_vr - FileNotFoundError: [WinError 2] The system cannot find the file specified
