### 2025-07-11 SWIFTSIM/swiftgalaxy#55 @blakeaw
And a few `ValueError`s:

> FAILED tests/test_halo_catalogues.py::TestHaloCatalogues::test_get_bound_only_extra_mask[soap] - ValueError: Unable to move link (name doesn't exist)
> FAILED tests/test_halo_catalogues.py::TestHaloCataloguesMulti::test_generate_extra_mask[soap] - ValueError: Unable to move link (name doesn't exist)
> FAILED tests/test_halo_catalogues.py::TestCaesarWithSWIFTGalaxy::test_incomplete_catalogue[halo] - ValueError: The truth value of an array with more than one element is ambiguous. Use a.any() or a.all()
