### 2025-07-11 SWIFTSIM/swiftgalaxy#55 @kyleaoman
Thanks for this. I think that a fix is probably straightforward, once a test fails to properly clean up after itself that will wreak havoc throughout the suite, so I think there's probably just a small issue with a file not being removed, as you guessed. Could you identify the first test that fails (chronologically, as the suite runs) and run it in isolation? If it fails in isolation the log of that failure would be valuable. If it passes then the failure will be due to some earlier successful test leaving something behind, and I'll need to chase it up more carefully.
