### 2025-07-11 SWIFTSIM/swiftgalaxy#55 @kyleaoman
Ah, and I've been idly wondering and vaguely searching for something like this for a while: https://docs.pytest.org/en/stable/how-to/tmp_path.html

I'll see if I can get the tests to stop messing each other up because of shared file names (from the fixtures).
