### 2025-07-11 SWIFTSIM/swiftgalaxy#55 @blakeaw
Parsing through the pytest output, it looks like this is the first failure listed:

> FAILED tests/test_coordinate_transformations.py::TestCopyingTransformations::test_invalid_coordinate_frame_from - PermissionError: [WinError 32] The process cannot access the file because it is being used by another process: 'demo_data\\toysnap.hdf5'

But the first error reported is:

> _ ERROR at teardown of TestAutoCoordinateTransformations.test_auto_recentering[gas-expected_xy0-expected_z0] _
>
> request = <SubRequest 'sg' for <Function test_auto_recentering[gas-expected_xy0-expected_z0]>>
>
>     @pytest.fixture(scope="function")
>     def sg(request):
>         _create_toysnap()
>
>         yield SWIFTGalaxy(
>             _toysnap_filename,
>             ToyHF(),
>             transforms_like_coordinates={"coordinates", "extra_coordinates"},
>             transforms_like_velocities={"velocities", "extra_velocities"},
>         )
>
> >       _remove_toysnap()
>
> conftest.py:79:
> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
> ..\swiftgalaxy\demo_data.py:1078: in _remove_toysnap
>     Path(snapfile).unlink(missing_ok=True)
> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
>
> self = WindowsPath('demo_data/toysnap.hdf5'), missing_ok = True
>
>     def unlink(self, missing_ok=False):
>         """
>         Remove this file or link.
>         If the path is a directory, use rmdir() instead.
>         """
>         try:
> >           self._accessor.unlink(self)
> E           PermissionError: [WinError 32] The process cannot access the file because it is being used by another process: 'demo_data\\toysnap.hdf5'

If I run just that one, `pytest --cov --cov-branch .\test_coordinate_transformations.py`, it gives the same error.
