### 2025-07-11 SWIFTSIM/swiftgalaxy#55 @blakeaw
Here's the full capture of running the tests in case it's helpful:

[output_tests.txt](https://github.com/user-attachments/files/21189258/output_tests.txt)
