### 2025-07-11 SWIFTSIM/swiftgalaxy#56 @blakeaw
pyOpenSci review - JOSS paper.md
Hi, this issue is part of this package's pyOpenSci review at https://github.com/pyOpenSci/software-submission/issues/244

**Severity:** Moderate

### Description

I see in the publication options (https://github.com/pyOpenSci/software-submission/issues/244#issue-3088374714) that the package is selected for JOSS submission. However, I don't see a JOSS `paper.md` associated with the repository.

Opening an issue here just in case the `paper.md` fell off the radar, and also, to help me keep track of it.
