### 2025-07-11 SWIFTSIM/swiftgalaxy#57 @blakeaw
pyOpenSci review - No API documentation for the demo_data module
Hi, this issue is part of this package's pyOpenSci review at https://github.com/pyOpenSci/software-submission/issues/244

**Severity:** Minor

### Description

In the docs, the API documentation entry for the `demo_data` module is empty (https://swiftgalaxy.readthedocs.io/en/latest/modules/swiftgalaxy.demo_data.html).

It'd probably be good to include at least some info about the demo data helpers (maybe just a list of available helpers/datasets) in the module-level docstring, and perhaps a link to the Example data page (https://swiftgalaxy.readthedocs.io/en/latest/demonstration_data/index.html).
