### 2025-07-16 acc78aa
Add unit tests for LazyMask class.
