+++ tests/conftest.py
@@ -12,6 +12,7 @@ from swiftgalaxy import (
     Standalone,
     MaskCollection,
 )
+from swiftgalaxy.masks import LazyMask
 from swiftgalaxy.demo_data import (
     _create_toysnap,
     _remove_toysnap,
@@ -874,3 +875,26 @@ def hf_multi_zerotarget(request):
                 scale_exponent=1,
             ),
         )
+
+
+@pytest.fixture(scope="function")
+def lm():
+
+    def mf(mask_loaded_data=None):
+        """
+        A simple mask function.
+
+        Parameters
+        ----------
+        mask_loaded_data : bool
+            For toggling masking data as it is loaded. Ignored in this test case.
+
+        Returns
+        -------
+        out : ndarray
+            A simple mask array.
+        """
+        return np.ones(10, dtype=bool)
+
+    lm = LazyMask(mask_function=mf)
+    yield lm
