+++ tests/test_masking.py
@@ -3,6 +3,7 @@ Tests for applying masks to swiftgalaxy, datasets and named columns.
 """
 
 import pytest
+from copy import copy, deepcopy
 import numpy as np
 from unyt.testing import assert_allclose_units
 from swiftgalaxy import SWIFTGalaxy, MaskCollection
@@ -17,6 +18,7 @@ from swiftgalaxy.demo_data import (
     _n_s_all,
     _n_bh_all,
 )
+from swiftgalaxy.masks import LazyMask
 
 abstol_nd = 1.0e-4
 reltol_nd = 1.0e-4
@@ -254,3 +256,196 @@ class TestMultiModeMask:
             RuntimeError, match="Tried to mask catalogue without mask index!"
         ):
             caesar_multi._mask_catalogue()
+
+
+class TestLazyMask:
+    """
+    Unit tests for the LazyMask class itself.
+    """
+
+    def test_init_lazy(self, lm):
+        """
+        Check that initializing in lazy mode works.
+        """
+        assert lm._evaluated is False
+        assert lm._mask_function is not None  # also implies it exists
+        assert not hasattr(lm, "_mask")
+
+    def test_init_concrete(self):
+        """
+        Check that initializing with a concrete mask works.
+        """
+        m = np.ones(10, dtype=bool)
+        lm = LazyMask(mask=m)
+        assert lm._evaluated
+        assert lm._mask is m
+        assert lm._mask_function is None
+
+    def test_init_none(self):
+        """
+        Check that initializing with ``None`` works (i.e. no mask).
+        """
+        lm = LazyMask(mask=None)
+        assert lm._evaluated
+        assert lm._mask is None
+        assert not hasattr(lm, "_mask_function")
+
+    def test_trigger_eval(self, lm):
+        """
+        Check that accessing mask triggers evaluation if lazy.
+        """
+        assert lm._evaluated is False
+        assert not hasattr(lm, "_mask")
+        assert (lm.mask == lm._mask_function()).all()
+        assert lm._evaluated
+        assert (lm._mask == lm._mask_function()).all()
+
+    def test_access_not_lazy(self):
+        """
+        Check that accessing the mask works for a non-lazy mask.
+        """
+        m = np.ones(10, dtype=bool)
+        lm = LazyMask(mask=m)
+        assert lm._evaluated
+        assert (lm.mask == m).all()
+
+    def test_manual_trigger_eval(self, lm):
+        """
+        Check that accessing mask triggers evaluation if lazy.
+        """
+
+        assert lm._evaluated is False
+        assert not hasattr(lm, "_mask")
+        lm._evaluate()
+        assert lm._evaluated
+        assert (lm.mask == lm._mask_function()).all()
+        assert (lm._mask == lm._mask_function()).all()
+
+    def test_trigger_eval_once_only(self):
+        """
+        Check that we can't trigger mask evaluation repeatedly.
+        """
+
+        class MF(object):
+            """
+            A simple class that behaves as a mask generator and counts how many times its
+            ``__call__`` is called.
+            """
+
+            call_counter: int = 0
+
+            def __call__(self, mask_loaded_data=None):
+                """
+                Call the class to behave like a simple mask function.
+
+                Parameters
+                ----------
+                mask_loaded_data : bool
+                    For toggling masking data as it is loaded. Ignored in this test case.
+
+                Returns
+                -------
+                out : ndarray
+                    A simple mask array.
+                """
+
+                self.call_counter += 1
+                return np.ones(10, dtype=bool)
+
+        mf = MF()
+        lm = LazyMask(mask_function=mf)
[…]
