### 2025-07-16 caaf103
Skip web example tests while server is down.
