+++ tests/test_examples.py
@@ -16,7 +16,6 @@ from swiftgalaxy.demo_data import (
 )
 
 
-@pytest.mark.skip
 class TestWebExampleData:
 
     def test_str(self):
