### 2025-07-16 10d112e
Update docstrings.
