### 2025-07-18 3791ba1
Tests for lazy masking for other catalogue types, removed _get_preload_fields entirely.
