### 2025-07-21 032f725
Implement lazy masks in the ToyHF.
