### 2025-07-21 9bbca3c
Evaluate masks lazily (#49)
* Basic implementation of lazy masking complete.

* Refactor caesar masks, pass linting.

* Give responsibility for masking data already in memory to SWIFTGalaxy.

* Avoid infinite recursion in lazy mask lookup.

* Catch all the cases where we need to mask already-loaded data.

* Fix a bug where None could be applied as a mask.

* Move lazy masking logic to a helper class.

* Don't mask mask construction data when preparing server masks.

* Add unit tests for LazyMask class.

* Skip web example tests while server is down.

* Clean notebook run.

* Fixes from coverag
[…]
