+++ tests/test_halo_catalogues.py
@@ -179,11 +179,11 @@ class TestHaloCatalogues:
         for particle_type in _present_particle_types.values():
             if expected_shape[particle_type] is not None:
                 assert (
-                    getattr(generated_extra_mask, particle_type).shape
+                    getattr(generated_extra_mask, particle_type).mask.shape
                     == expected_shape[particle_type]
                 )
                 assert (
-                    getattr(generated_extra_mask, particle_type).sum()
+                    getattr(generated_extra_mask, particle_type).mask.sum()
                     == dict(
                         gas=_n_g_1,
                         dark_matter=_n_dm_1,
@@ -224,7 +224,7 @@ class TestHaloCatalogues:
                 )
             else:
                 assert (
-                    getattr(generated_extra_mask, particle_type).sum()
+                    getattr(generated_extra_mask, particle_type).mask.sum()
                     == dict(gas=100, dark_matter=None, stars=100, black_holes=_n_bh_1)[
                         particle_type
                     ]
@@ -413,11 +413,11 @@ class TestHaloCataloguesMulti:
         for particle_type in _present_particle_types.values():
             if expected_shape[particle_type] is not None:
                 assert (
-                    getattr(generated_extra_mask, particle_type).shape
+                    getattr(generated_extra_mask, particle_type).mask.shape
                     == expected_shape[particle_type]
                 )
                 assert (
-                    getattr(generated_extra_mask, particle_type).sum()
+                    getattr(generated_extra_mask, particle_type).mask.sum()
                     == dict(
                         gas=_n_g_1,
                         dark_matter=_n_dm_1,
@@ -658,9 +658,82 @@ class TestVelociraptorWithSWIFTGalaxy:
             )
             for ptype in _present_particle_types.values():
                 assert np.all(
-                    getattr(sg_from_sgs._extra_mask, ptype)
-                    == getattr(sg._extra_mask, ptype)
+                    getattr(sg_from_sgs._extra_mask, ptype).mask
+                    == getattr(sg._extra_mask, ptype).mask
                 )
+
+    def test_lazy_masking_sg(self, sg_vr):
+        """
+        Check that lazy masks are evaluated only when needed, loading and masking exactly
+        the expected data.
+        """
+        # check that masks aren't evaluated
+        for ptype in sg_vr.metadata.present_group_names:
+            if getattr(sg_vr._extra_mask, ptype)._mask_function is not None:
+                assert getattr(sg_vr._extra_mask, ptype)._evaluated is False
+            particle_dataset = getattr(sg_vr, ptype)._particle_dataset
+            for field_name in getattr(particle_dataset, "group_metadata").field_names:
+                assert getattr(particle_dataset, f"_{field_name}") is None
+        # trigger mask evaluation and application:
+        sg_vr.gas.masses
+        # now check that mask was evaluated and applied for gas:
+        assert sg_vr._extra_mask.gas._evaluated
+        assert sg_vr.gas._particle_dataset._particle_ids is not None
+        assert sg_vr.gas.particle_ids.shape == sg_vr.gas.masses.shape
+        assert sg_vr._extra_mask.gas._mask.shape != sg_vr.gas.masses.shape
+        # and other masks are still unevaluated, no other data loaded
+        for ptype in sg_vr.metadata.present_group_names:
+            if ptype != "gas":
+                if getattr(sg_vr._extra_mask, ptype)._mask_function is not None:
+                    assert getattr(sg_vr._extra_mask, ptype)._evaluated is False
+            particle_dataset = getattr(sg_vr, ptype)._particle_dataset
+            for field_name in getattr(particle_dataset, "group_metadata").field_names:
+                if ptype == "gas" and field_name in ("masses", "particle_ids"):
+                    continue
[…]
