+++ swiftgalaxy/reader.py
@@ -31,7 +31,7 @@ from swiftsimio.reader import (
 from swiftsimio.objects import cosmo_array
 from swiftsimio.masks import SWIFTMask
 from swiftgalaxy.halo_catalogues import _HaloCatalogue
-from swiftgalaxy.masks import MaskCollection
+from swiftgalaxy.masks import MaskCollection, LazyMask
 
 from typing import Union, Optional, Set, Callable
 
@@ -61,9 +61,9 @@ def _apply_box_wrap(
     current_transform : :class:`~numpy.ndarray`
         The currently active 4x4 transformation matrix.
 
-    offset_frac : :obj:`float`
+    offset_frac : :obj:`float`, default: ``0.5``
         The fraction of the box to offset by. The default it to wrap to [-Lbox/2, Lbox/2].
-        Setting to 0.0 instead wraps to [0, Lbox]. (Default: 0.5)
+        Setting to 0.0 instead wraps to [0, Lbox].
 
     Returns
     -------
@@ -823,10 +823,10 @@ class _SWIFTGroupDatasetHelper(__SWIFTGroupDataset):
                 self._swiftgalaxy._extra_mask, self._particle_dataset.group_name
             )
             if mask is not None:
-                return data[mask]
+                return data[mask.mask]
         return data
 
-    def _mask_dataset(self, mask: slice) -> None:
+    def _mask_dataset(self, mask: LazyMask) -> None:
         """
         Apply a mask to this data set.
 
@@ -834,7 +834,7 @@ class _SWIFTGroupDatasetHelper(__SWIFTGroupDataset):
 
         Parameters
         ----------
-        mask : :obj:`slice`
+        mask : :class:`~swiftgalaxy.masks.LazyMask`
             The mask to apply to all data arrays managed by this
             :class:`~swiftgalaxy.reader._SWIFTGroupDatasetHelper`.
         """
@@ -844,6 +844,12 @@ class _SWIFTGroupDatasetHelper(__SWIFTGroupDataset):
         particle_metadata = getattr(
             self._particle_dataset.metadata, f"{particle_name}_properties"
         )
+        # force old mask evaluation to ensure any data loaded during evaluation
+        # are in memory and have the old mask applied, if any:
+        old_mask = getattr(self._swiftgalaxy._extra_mask, particle_name)
+        if old_mask is not None:
+            old_mask._evaluate()
+        # apply the new mask to any data already in memory:
         for field_name in particle_metadata.field_names:
             if self._is_namedcolumns(field_name):
                 for named_column in getattr(self, field_name).named_columns:
@@ -857,10 +863,11 @@ class _SWIFTGroupDatasetHelper(__SWIFTGroupDataset):
                         setattr(
                             getattr(self, field_name),
                             named_column,
-                            getattr(getattr(self, field_name), named_column)[mask],
+                            getattr(getattr(self, field_name), named_column)[mask.mask],
                         )
             elif getattr(self._particle_dataset, f"_{field_name}") is not None:
-                setattr(self, field_name, getattr(self, field_name)[mask])
+                setattr(self, field_name, getattr(self, field_name)[mask.mask])
+        # also the derived coordinates, if any:
         self._mask_derived_coordinates(mask)
         if getattr(self._swiftgalaxy._extra_mask, particle_name) is None:
             setattr(self._swiftgalaxy._extra_mask, particle_name, mask)
@@ -875,13 +882,12 @@ class _SWIFTGroupDatasetHelper(__SWIFTGroupDataset):
                         particle_name
                     ]
                 )
-            old_mask = getattr(self._swiftgalaxy._extra_mask, particle_name)
             # need to convert to an integer mask to combine
[…]
