### 2025-07-26 563cdad
Move all test data except examples to tmp directories.
