+++ tests/test_halo_catalogues.py
@@ -13,6 +13,8 @@ from swiftgalaxy.demo_data import (
     _toycaesar_filename,
     _toysoap_virtual_snapshot_filename,
     _toysoap_membership_filebase,
+    _create_toysnap,
+    _remove_toysnap,
     _create_toycaesar,
     _remove_toycaesar,
     _n_g_1,
@@ -57,8 +59,8 @@ class TestHaloCatalogues:
         """
         # don't use sg fixture here, just need the snapshot file
         # so don't want overhead of a SWIFTGalaxy
-        spatial_mask = hf._get_spatial_mask(_toysnap_filename)
-        with h5py.File(_toysnap_filename, "r") as snap:
+        spatial_mask = hf._get_spatial_mask(toysnap["toysnap_filename"])
+        with h5py.File(toysnap["toysnap_filename"], "r") as snap:
             n_g_firstcell = snap["/Cells/Counts/PartType0"][0]
             n_dm_firstcell = snap["/Cells/Counts/PartType1"][0]
             n_s_firstcell = snap["/Cells/Counts/PartType4"][0]
@@ -107,9 +109,9 @@ class TestHaloCatalogues:
             scale_exponent=1,
         )
         hf.extra_mask = None
-        sg = SWIFTGalaxy(_toysnap_filename, hf)
+        sg = SWIFTGalaxy(toysnap["toysnap_filename"], hf)
         generated_spatial_mask = sg._spatial_mask
-        with h5py.File(_toysnap_filename, "r") as snap:
+        with h5py.File(toysnap["toysnap_filename"], "r") as snap:
             n_g_firstcell = snap["/Cells/Counts/PartType0"][0]
             n_dm_firstcell = snap["/Cells/Counts/PartType1"][0]
             n_s_firstcell = snap["/Cells/Counts/PartType4"][0]
@@ -131,7 +133,7 @@ class TestHaloCatalogues:
             np.array([[0, n_bh_firstcell], [n_bh_firstcell, _n_bh_1 + _n_bh_2]]),
         )
 
-    def test_get_bound_only_extra_mask(self, hf, toysnap_withfof):
+    def test_get_bound_only_extra_mask(self, tmp_path_factory, hf):
         """
         Check that bound_only extra mask has the right shape.
         """
@@ -140,36 +142,55 @@ class TestHaloCatalogues:
             from compression.make_virtual_snapshot import make_virtual_snapshot
             from compression.update_vds_paths import update_virtual_snapshot_paths
 
-            membership_filepattern = (
-                str(_toysoap_membership_filebase) + ".{file_nr}.hdf5"
+            tp = hf.soap_file.parent
+            toysnap_filename = tp / _toysnap_filename.name
+            _create_toysnap(snapfile=toysnap_filename, withfof=True)
+            membership_filepattern = tp / (
+                str(_toysoap_membership_filebase.name) + ".{file_nr}.hdf5"
+            )
+            toysoap_virtual_snapshot_filename = (
+                tp / _toysoap_virtual_snapshot_filename.name
             )
             make_virtual_snapshot(
-                _toysnap_filename,
-                membership_filepattern,
-                _toysoap_virtual_snapshot_filename,
+                toysnap_filename,
+                str(membership_filepattern),
+                toysoap_virtual_snapshot_filename,
                 0,  # snapshot number, not used since no pattern in filenames
             )
-            abs_snapshot_dir = Path(_toysnap_filename).parent.absolute()
+            abs_snapshot_dir = Path(toysnap_filename).parent.absolute()
             abs_membership_dir = Path(
-                membership_filepattern.format(file_nr=0)
+                str(membership_filepattern).format(file_nr=0)
             ).parent.absolute()
-            abs_output_dir = Path(_toysoap_virtual_snapshot_filename).parent.absolute()
[…]
