+++ swiftgalaxy/demo_data.py
@@ -1147,8 +1147,8 @@ def _create_toyvr(filebase: Union[str, Path] = _toyvr_filebase) -> None:
         The base name for catalogue files (several files ``base.properties``,
         ``base.catalog_groups``, etc. will be created).
     """
-    if not Path(f"{_toyvr_filebase}.properties").is_file():
-        with h5py.File(f"{_toyvr_filebase}.properties", "w") as f:
+    if not Path(f"{filebase}.properties").is_file():
+        with h5py.File(f"{filebase}.properties", "w") as f:
             f.create_group("SimulationInfo")
             f["SimulationInfo"].attrs["ScaleFactor"] = 1.0
             f["SimulationInfo"].attrs["Cosmological_Sim"] = 1
@@ -1320,8 +1320,8 @@ def _create_toyvr(filebase: Union[str, Path] = _toyvr_filebase) -> None:
             f["numSubStruct"].attrs["Dimension_Mass"] = 0.0
             f["numSubStruct"].attrs["Dimension_Time"] = 0.0
             f["numSubStruct"].attrs["Dimension_Velocity"] = 0.0
-    if not Path(f"{_toyvr_filebase}.catalog_groups").is_file():
-        with h5py.File(f"{_toyvr_filebase}.catalog_groups", "w") as f:
+    if not Path(f"{filebase}.catalog_groups").is_file():
+        with h5py.File(f"{filebase}.catalog_groups", "w") as f:
             f.create_dataset("File_id", data=np.array([0, 0], dtype=int))
             f.create_dataset(
                 "Group_Size",
@@ -1348,8 +1348,8 @@ def _create_toyvr(filebase: Union[str, Path] = _toyvr_filebase) -> None:
             )
             f.create_dataset("Parent_halo_ID", data=np.array([-1, -1], dtype=int))
             f.create_dataset("Total_num_of_groups", data=np.array([2], dtype=int))
-    if not Path(f"{_toyvr_filebase}.catalog_particles").is_file():
-        with h5py.File(f"{_toyvr_filebase}.catalog_particles", "w") as f:
+    if not Path(f"{filebase}.catalog_particles").is_file():
+        with h5py.File(f"{filebase}.catalog_particles", "w") as f:
             f.create_dataset("File_id", data=np.array([0], dtype=int))
             f.create_dataset("Num_of_files", data=np.array([1], dtype=int))
             f.create_dataset(
@@ -1425,8 +1425,8 @@ def _create_toyvr(filebase: Union[str, Path] = _toyvr_filebase) -> None:
                     dtype=int,
                 ),
             )
-    if not Path(f"{_toyvr_filebase}.catalog_particles.unbound").is_file():
-        with h5py.File(f"{_toyvr_filebase}.catalog_particles.unbound", "w") as f:
+    if not Path(f"{filebase}.catalog_particles.unbound").is_file():
+        with h5py.File(f"{filebase}.catalog_particles.unbound", "w") as f:
             f.create_dataset("File_id", data=np.array([0], dtype=int))
             f.create_dataset("Num_of_files", data=np.array([1], dtype=int))
             f.create_dataset(
@@ -1452,8 +1452,8 @@ def _create_toyvr(filebase: Union[str, Path] = _toyvr_filebase) -> None:
                 "Total_num_of_particles_in_all_groups",
                 data=np.array([_n_g_b + _n_dm_b], dtype=int),
             )
-    if not Path(f"{_toyvr_filebase}.catalog_parttypes").is_file():
-        with h5py.File(f"{_toyvr_filebase}.catalog_parttypes", "w") as f:
+    if not Path(f"{filebase}.catalog_parttypes").is_file():
[…]
