### 2025-07-26 4831a71
Move web and generated examples to tmp directories in testing.
