+++ tests/test_examples.py
@@ -16,21 +16,33 @@ from swiftgalaxy.demo_data import (
 )
 
 
+@pytest.fixture(scope="session")
+def generated_examples_tmpdir(tmp_path_factory):
+    generated_examples._demo_data_dir = tmp_path_factory.mktemp("demo_data")
+    return generated_examples
+
+
+@pytest.fixture(scope="session")
+def web_examples_tmpdir(tmp_path_factory):
+    web_examples._demo_data_dir = tmp_path_factory.mktemp("demo_data")
+    return web_examples
+
+
 class TestWebExampleData:
 
-    def test_str(self):
+    def test_str(self, web_examples_tmpdir):
         """
         Check that string representation lists available examples.
         """
-        assert "snapshot" in str(web_examples)
+        assert "snapshot" in str(web_examples_tmpdir)
 
-    def test_snapshot(self):
+    def test_snapshot(self, web_examples_tmpdir):
         """
         Check that we can create a swiftgalaxy, retrieving a sample snapshot file.
         """
-        sm = mask(web_examples.snapshot)  # just to get an interface to metadata
+        sm = mask(web_examples_tmpdir.snapshot)  # just to get an interface to metadata
         SWIFTGalaxy(
-            web_examples.snapshot,
+            web_examples_tmpdir.snapshot,
             Standalone(
                 centre=cosmo_array(
                     [2, 2, 2],
@@ -57,115 +69,120 @@ class TestWebExampleData:
         )
 
     @pytest.mark.parametrize("group_type", ["halo", "galaxy"])
-    def test_caesar(self, group_type):
+    def test_caesar(self, web_examples_tmpdir, group_type):
         """
         Check that we can create a swiftgalaxy, retrieving a sample snapshot file and
         caesar catalogue.
         """
         SWIFTGalaxy(
-            web_examples.snapshot,
-            Caesar(web_examples.caesar, group_type=group_type, group_index=0),
+            web_examples_tmpdir.snapshot,
+            Caesar(web_examples_tmpdir.caesar, group_type=group_type, group_index=0),
         )
 
-    def test_soap(self):
+    def test_soap(self, web_examples_tmpdir):
         """
         Check that we can create a swiftgalaxy, retrieving a sample virtual snapshot
         file and a soap catalogue.
         """
         SWIFTGalaxy(
-            web_examples.virtual_snapshot, SOAP(web_examples.soap, soap_index=0)
+            web_examples_tmpdir.virtual_snapshot,
+            SOAP(web_examples_tmpdir.soap, soap_index=0),
         )
 
-    def test_vr(self):
+    def test_vr(self, web_examples_tmpdir):
         """
         Check that we can create a swiftgalaxy, retrieving a sample snapshot file
         and velociraptor catalogue.
         """
         SWIFTGalaxy(
-            web_examples.snapshot, Velociraptor(web_examples.velociraptor, halo_index=0)
+            web_examples_tmpdir.snapshot,
+            Velociraptor(web_examples_tmpdir.velociraptor, halo_index=0),
         )
 
-    def test_remove(self):
+    def test_remove(self, web_examples_tmpdir):
         """
         Check that example data files get cleaned up on request.
         """
         # download all the example data (if not present)
-        for example in web_examples.available_examples.keys():
-            getattr(web_examples, example)
-        web_examples.remove()
-        for absent_files in web_examples.available_examples.values():
+        for example in web_examples_tmpdir.available_examples.keys():
+            getattr(web_examples_tmpdir, example)
+        web_examples_tmpdir.remove()
+        for absent_files in web_examples_tmpdir.available_examples.values():
             for absent_file in absent_files:
                 assert not Path(absent_file).is_file()
 
 
 class TestGeneratedExampleData:
 
[…]
