### 2025-07-26 59a9fe9
Fix a couple of last bugs in tests.
