### 2025-07-28 9c8346a
Remove demo data directory along with files, if empty (enables notebook cleaning up after itself in tests).
