### 2025-07-28 0e3abb4
Move fixtures to conftest.
