### 2025-07-28 86da2b7
Try switching on parallel tests in CI.
