+++ .github/workflows/lint_and_test.yml
@@ -65,13 +65,13 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install flake8 pytest
+        pip install flake8 pytest pytest-xdist
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
         if [ -f optional_requirements.txt ]; then pip install -r optional_requirements.txt; fi
         pip install -e .
     - name: Test with pytest
       run: |
-        pytest
+        pytest -n 2
 
   tests-and-coverage:
     needs: lint
@@ -97,13 +97,13 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install flake8 pytest pytest-cov
+        pip install flake8 pytest pytest-cov pytest-xdist
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
         if [ -f optional_requirements.txt ]; then pip install -r optional_requirements.txt; fi
         pip install -e .
     - name: Test with pytest including coverage report
       run: |
-        pytest --cov --cov-branch --cov-report=xml
+        pytest -n 2 --cov --cov-branch --cov-report=xml
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v5
       env:
