### 2025-07-29 f42630b
Use non-default directory for notebook test data.
