### 2025-07-29 a36473b
Use temporary directories for test data (#60)
* Move all test data except examples to tmp directories.

* Move web and generated examples to tmp directories in testing.

* Fix a couple of last bugs in tests.

* Remove demo data directory along with files, if empty (enables notebook cleaning up after itself in tests).

* Move fixtures to conftest.

* Try switching on parallel tests in CI.

* Use non-default directory for notebook test data.

* Ensure coverage.
