+++ tests/conftest.py
@@ -25,6 +25,7 @@ from swiftgalaxy.demo_data import (
     ToyHF,
     _toysnap_filename,
     _toysoap_filename,
+    _toysoap_membership_filebase,
     _toysoap_virtual_snapshot_filename,
     _toyvr_filebase,
     _toycaesar_filename,
@@ -34,58 +35,94 @@ from swiftgalaxy.demo_data import (
     _centre_2,
     _vcentre_1,
     _vcentre_2,
+    generated_examples,
+    web_examples,
 )
 
 hfs = ("vr", "caesar_halo", "caesar_galaxy", "sa", "soap")
 
 
 @pytest.fixture(scope="function")
-def toysnap():
-    _create_toysnap()
+def toysnap(tmp_path_factory):
+    toysnap_filename = (
+        tmp_path_factory.mktemp(_toysnap_filename.parent) / _toysnap_filename.name
+    )
+    _create_toysnap(snapfile=toysnap_filename)
 
-    yield
+    yield {"toysnap_filename": toysnap_filename}
 
-    _remove_toysnap()
-
-
-@pytest.fixture(scope="function")
-def toysnap_withfof():
-    _create_toysnap(withfof=True)
-
-    yield
-
-    _remove_toysnap()
+    _remove_toysnap(snapfile=toysnap_filename)
 
 
 @pytest.fixture(scope="function")
-def toysoap_with_virtual_snapshot():
-    _create_toysoap(create_virtual_snapshot=True)
+def toysnap_withfof(tmp_path_factory):
+    toysnap_filename = (
+        tmp_path_factory.mktemp(_toysnap_filename.parent) / _toysnap_filename.name
+    )
+    _create_toysnap(snapfile=toysnap_filename, withfof=True)
 
-    yield
+    yield {"toysnap_filename": toysnap_filename}
 
-    _remove_toysoap()
+    _remove_toysnap(snapfile=toysnap_filename)
 
 
 @pytest.fixture(scope="function")
-def sg(request):
-    _create_toysnap()
+def toysoap_with_virtual_snapshot(tmp_path_factory):
+    tp = tmp_path_factory.mktemp(_toysoap_filename.parent)
+    toysoap_filename = tp / _toysoap_filename.name
+    membership_filebase = tp / _toysoap_membership_filebase.name
+    toysnap_filename = tp / _toysnap_filename.name
+    toysoap_virtual_snapshot_filename = tp / _toysoap_virtual_snapshot_filename.name
+    _create_toysnap(toysnap_filename, withfof=True)
+    _create_toysoap(
+        filename=toysoap_filename,
+        membership_filebase=membership_filebase,
+        create_virtual_snapshot=True,
+        create_virtual_snapshot_from=toysnap_filename,
+        virtual_snapshot_filename=toysoap_virtual_snapshot_filename,
+    )
+
+    yield {
+        "toysoap_filename": toysoap_filename,
+        "membership_filebase": membership_filebase,
+        "toysnap_filename": toysnap_filename,
+        "toysoap_virtual_snapshot_filename": toysoap_virtual_snapshot_filename,
+    }
+
+    _remove_toysoap(
+        filename=toysoap_filename,
+        membership_filebase=membership_filebase,
+        virtual_snapshot_filename=toysoap_virtual_snapshot_filename,
+    )
+    _remove_toysnap(snapfile=toysnap_filename)
+
+
+@pytest.fixture(scope="function")
+def sg(request, tmp_path_factory):
+    toysnap_filename = (
+        tmp_path_factory.mktemp(_toysnap_filename.parent) / _toysnap_filename.name
+    )
+    _create_toysnap(snapfile=toysnap_filename)
 
     yield SWIFTGalaxy(
-        _toysnap_filename,
[…]
