### 2025-07-29 SWIFTSIM/swiftgalaxy#53 @kyleaoman
Caesar does seem to declare `Cython` as a dependency for both building and installing in its `setup.py`, so I think that the issue is (probably?) the use of a legacy build framework. I've opened a `caesar` issue, will see if they decide to address it promptly, else could try to do something on this side - hints in the docs is probably the only option, don't really want a direct `Cython` dependency.
