### 2025-07-29 SWIFTSIM/swiftgalaxy#53 @dnarayanan
hi all - thanks for posting the issue in caesar!   i can see from this thread the issues that the current build system is causing, though i have to admit relative inexperience when it comes to package management and would very much welcome either suggestions, or of course even better, suggestions via a PR :)
