### 2025-07-29 b2cf966
Flesh out docstrings for demo_data module.
