### 2025-07-29 SWIFTSIM/swiftgalaxy#53 @kyleaoman
@dnarayanan could you try following the suggestions in https://github.com/pypa/pip/issues/6334, and perhaps find a mature package with a cython dependency and see how they do their build? I haven't packaged a cython-dependent package before so I'd be trying to stumble towards a solution myself, but those would be my starting points...

If you're still stuck after that drop another message here and maybe I can try to help?
