### 2025-08-04 1a0e3bb
Better opening to README (#62)
