### 2025-08-04 SWIFTSIM/swiftgalaxy#53 @kyleaoman
Have just tested more carefully and in a clean/blank environment I can install both with caesar's `master` and `new_installer` branches:
`pip install git+https://github.com/dnarayanan/caesar.git`
`pip install git+https://github.com/dnarayanan/caesar.git@new_installer`

So my success with the new branch is not a particularly discriminatory test of success. Maybe @blakeaw can see if that branch fixes the error on his setup?
