### 2025-08-04 68f2e73
Make numpydoc command platform-independent. (#63)
