### 2025-08-04 b3b7772
Add a JOSS paper draft.
