+++ bibliography.bib
@@ -1,0 +1,147 @@
+@ARTICLE{Borrow2020,
+       author = {{Borrow}, Josh and {Borrisov}, Alexei},
+        title = "{swiftsimio: A Python library for reading SWIFT data}",
+      journal = {The Journal of Open Source Software},
+     keywords = {Python, cosmology, i/o, simulations, astronomy},
+         year = 2020,
+        month = aug,
+       volume = {5},
+       number = {52},
+          eid = {2430},
+        pages = {2430},
+          doi = {10.21105/joss.02430},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020JOSS....5.2430B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Schaller2024,
+       author = {{Schaller}, Matthieu and {Borrow}, Josh and {Draper}, Peter W. and {Ivkovic}, Mladen and {McAlpine}, Stuart and {Vandenbroucke}, Bert and {Bah{\'e}}, Yannick and {Chaikin}, Evgenii and {Chalk}, Aidan B.~G. and {Chan}, Tsang Keung and {Correa}, Camila and {van Daalen}, Marcel and {Elbers}, Willem and {Gonnet}, Pedro and {Hausammann}, Lo{\"\i}c and {Helly}, John and {Hu{\v{s}}ko}, Filip and {Kegerreis}, Jacob A. and {Nobels}, Folkert S.~J. and {Ploeckinger}, Sylvia and {Revaz}, Yves and {Roper}, William J. and {Ruiz-Bonilla}, Sergio and {Sandnes}, Thomas D. and {Uyttenhove}, Yolan and {Willis}, James S. and {Xiang}, Zhen},
+        title = "{SWIFT: A modern highly-parallel gravity and smoothed particle hydrodynamics solver for astrophysical and cosmological applications}",
+      journal = {\mnras},
+     keywords = {software: simulations, methods: numerical, software: public release, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Astrophysics of Galaxies, Computer Science - Distributed, Parallel, and Cluster Computing},
+         year = 2024,
+        month = may,
+       volume = {530},
+       number = {2},
+        pages = {2378-2419},
+          doi = {10.1093/mnras/stae922},
+archivePrefix = {arXiv},
+       eprint = {2305.13380},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024MNRAS.530.2378S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Elahi2019,
+       author = {{Elahi}, Pascal J. and {Ca{\~n}as}, Rodrigo and {Poulton}, Rhys J.~J. and {Tobar}, Rodrigo J. and {Willis}, James S. and {Lagos}, Claudia del P. and {Power}, Chris and {Robotham}, Aaron S.~G.},
+        title = "{Hunting for galaxies and halos in simulations with VELOCIraptor}",
+      journal = {\pasa},
+     keywords = {dark matter, galaxies: evolution, galaxies: halos, methods: numerical, Astrophysics - Cosmology and Nongalactic Astrophysics},
+         year = 2019,
+        month = may,
+       volume = {36},
+          eid = {e021},
+        pages = {e021},
+          doi = {10.1017/pasa.2019.12},
+archivePrefix = {arXiv},
+       eprint = {1902.01010},
+ primaryClass = {astro-ph.CO},
[…]
