+++ paper.md
@@ -38,7 +38,7 @@ The collection of standard cosmological hydrodynamical simulation data products 
 
 SWIFTGalaxy extends SWIFTSimIO by implementing the workflow outlined above supplemented with additional features including coordinate transformations, data array masking and efficient iteration over galaxies; all of the SWIFTSimIO features, including use of the visualisation tools, are inherited by SWIFTGalaxy. This is intended to avoid duplication of effort by users who would otherwise each need to implement the same steps independently, and helps to ensure that the implementation provided is optimal. It also helps to avoid common errors, such as applying a coordinate transformation to one particle type (e.g. rotation of the dark matter) but forgetting to apply it to others (e.g. the gas remains unrotated) by enforcing a consistent data state where applicable. A core design principle of SWIFTGalaxy is that only operations with an unambiguous implementation are included - in other words, SWIFTGalaxy tries to avoid making any decisions for its users. This is a key difference when compared to other packages serving this purpose, such as pynbody [@Pontzen2013] and yt [@Turk2011] that can also be used with SWIFT data sets. Other packages are also less tailored to SWIFT and therefore less able to take advantage of the detailed structure of the data layout on disk and information available as metadata.
 
-SWIFTGalaxy is hosted on [GitHub](https://github.com/SWIFTSIM/swiftgalaxy) and has documentation available through [ReadTheDocs](https://swiftgalaxy.readthedocs.io).
+SWIFTGalaxy is hosted on [GitHub](https://github.com/SWIFTSIM/swiftgalaxy) and has documentation available through [ReadTheDocs](https://swiftgalaxy.readthedocs.io). The first research article using the code has recently appeared [@Trayford2025]; many more projects using the code are currently ongoing, especially in the context of the COLIBRE simulations project (Schaye et al., in preparation; Chaikin et al., in preparation).
 
 # Acknowledgements
 
