+++ bibliography.bib
@@ -145,3 +145,20 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Trayford2025,
+       author = {{Trayford}, James W. and {Schaye}, Joop and {Correa}, Camila and {Ploeckinger}, Sylvia and {Richings}, Alexander J. and {Chaikin}, Evgenii and {Schaller}, Matthieu and {Benitez-Llambay}, Alejandro and {Frenk}, Carlos and {Husko}, Filip},
+        title = "{Modelling the evolution and influence of dust in cosmological simulations that include the cold phase of the interstellar medium}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics of Galaxies},
+         year = 2025,
+        month = may,
+          eid = {arXiv:2505.13056},
+        pages = {arXiv:2505.13056},
+          doi = {10.48550/arXiv.2505.13056},
+archivePrefix = {arXiv},
+       eprint = {2505.13056},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2025arXiv250513056T},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
