### 2025-08-04 95ce3c9
Move supported catalogue formats out of summary section.
