+++ paper.md
@@ -22,7 +22,7 @@ bibliography: bibliography.bib
 
 # Summary
 
-SWIFTGalaxy is an open-source astrophysics module that extends SWIFTSimIO [@Borrow2020] tailored to analyses of particles belonging to individual galaxies simulated with SWIFT [@Schaller2024]. It inherits from and extends the functionality of the SWIFTSimIO's SWIFTDataset class. It understands the content of halo catalogues (supported: Velociraptor [@Elahi2019], [Caesar](https://github.com/dnarayanan/caesar), SOAP [@McGibbon2025]) and therefore which particles belong to a galaxy or other group of particles, and its integrated properties. The particles occupy a coordinate frame that is enforced to be consistent, such that particles loaded on-the-fly will match e.g. rotations and translations of particles already in memory. Intuitive masking of particle datasets is also enabled. Utilities to make working in cylindrical and spherical coordinate systems more convenient are also provided. Finally, tools to iterate efficiently over multiple galaxies are provided.
+SWIFTGalaxy is an open-source astrophysics module that extends SWIFTSimIO [@Borrow2020] tailored to analyses of particles belonging to individual galaxies simulated with SWIFT [@Schaller2024]. It inherits from and extends the functionality of the SWIFTSimIO's SWIFTDataset class. It understands the content of halo catalogues and therefore which particles belong to a galaxy or other group of particles, and its integrated properties. The particles occupy a coordinate frame that is enforced to be consistent, such that particles loaded on-the-fly will match e.g. rotations and translations of particles already in memory. Intuitive masking of particle datasets is also enabled. Utilities to make working in cylindrical and spherical coordinate systems more convenient are also provided. Finally, tools to iterate efficiently over multiple galaxies are provided.
 
 # Background
 
@@ -36,7 +36,11 @@ For many types of analysis the snapshot files alone are insufficient: knowledge 
 
 The collection of standard cosmological hydrodynamical simulation data products - snapshots of particles, halo membership information and halo catalgues - leads to a generic workflow to begin analysis of individual galaxies in such simulations. First, an object of interest is identified in the halo catalogue. Then, the halo membership information is queried to locate its member particles. Finally, the particles are loaded from the snapshot to proceed with analysis. The SWIFT community already maintains the SWIFTSimIO tool [@Borrow2020] that supports reading in SWIFT snapshot files and SOAP catalogue files. It uses metadata in the files to annotate data arrays with physical units and relevant cosmological information. It is also able to efficiently select spatial sub-regions of a simulation by taking advantage of the fact the particles in cells of a "top-level cell" grid covering the simulation domain are stored contiguously. Finally, it includes some data visualisation tools.
 
[…]
