### 2025-08-04 f9846b9
Hard-code journal names.
