+++ bibliography.bib
@@ -17,7 +17,7 @@
 @ARTICLE{Schaller2024,
        author = {{Schaller}, Matthieu and {Borrow}, Josh and {Draper}, Peter W. and {Ivkovic}, Mladen and {McAlpine}, Stuart and {Vandenbroucke}, Bert and {Bah{\'e}}, Yannick and {Chaikin}, Evgenii and {Chalk}, Aidan B.~G. and {Chan}, Tsang Keung and {Correa}, Camila and {van Daalen}, Marcel and {Elbers}, Willem and {Gonnet}, Pedro and {Hausammann}, Lo{\"\i}c and {Helly}, John and {Hu{\v{s}}ko}, Filip and {Kegerreis}, Jacob A. and {Nobels}, Folkert S.~J. and {Ploeckinger}, Sylvia and {Revaz}, Yves and {Roper}, William J. and {Ruiz-Bonilla}, Sergio and {Sandnes}, Thomas D. and {Uyttenhove}, Yolan and {Willis}, James S. and {Xiang}, Zhen},
         title = "{SWIFT: A modern highly-parallel gravity and smoothed particle hydrodynamics solver for astrophysical and cosmological applications}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {software: simulations, methods: numerical, software: public release, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Astrophysics of Galaxies, Computer Science - Distributed, Parallel, and Cluster Computing},
          year = 2024,
         month = may,
@@ -35,7 +35,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Elahi2019,
        author = {{Elahi}, Pascal J. and {Ca{\~n}as}, Rodrigo and {Poulton}, Rhys J.~J. and {Tobar}, Rodrigo J. and {Willis}, James S. and {Lagos}, Claudia del P. and {Power}, Chris and {Robotham}, Aaron S.~G.},
         title = "{Hunting for galaxies and halos in simulations with VELOCIraptor}",
-      journal = {\pasa},
+      journal = {Publications of the Astronomical Society of Australia},
      keywords = {dark matter, galaxies: evolution, galaxies: halos, methods: numerical, Astrophysics - Cosmology and Nongalactic Astrophysics},
          year = 2019,
         month = may,
@@ -86,7 +86,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Han2018,
        author = {{Han}, Jiaxin and {Cole}, Shaun and {Frenk}, Carlos S. and {Benitez-Llambay}, Alejandro and {Helly}, John},
         title = "{HBT+: an improved code for finding subhaloes and building merger trees in cosmological simulations}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {gravitational lensing: strong, methods: numerical, galaxies: haloes, dark matter, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Astrophysics of Galaxies},
          year = 2018,
         month = feb,
@@ -115,7 +115,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Turn2011,
        author = {{Turk}, Matthew J. and {Smith}, Britton D. and {Oishi}, Jeffrey S. and {Skory}, Stephen and {Skillman}, Samuel W. and {Abel}, Tom and {Norman}, Michael L.},
         title = "{yt: A Multi-code Analysis Toolkit for Astrophysical Simulation Data}",
-      journal = {\apjs},
+      journal = {The Astrophysical Journal Supplements},
[…]
