+++ bibliography.bib
@@ -101,7 +101,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@software{@Pontzen2013,
+@software{Pontzen2013,
        author = {{Pontzen}, Andrew and {Ro{\v{s}}kar}, Rok and {Stinson}, Greg and {Woods}, Rory},
         title = "{pynbody: N-Body/SPH analysis for python}",
  howpublished = {Astrophysics Source Code Library, record ascl:1305.002},
@@ -112,7 +112,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{Turn2011,
+@ARTICLE{Turk2011,
        author = {{Turk}, Matthew J. and {Smith}, Britton D. and {Oishi}, Jeffrey S. and {Skory}, Stephen and {Skillman}, Samuel W. and {Abel}, Tom and {Norman}, Michael L.},
         title = "{yt: A Multi-code Analysis Toolkit for Astrophysical Simulation Data}",
       journal = {The Astrophysical Journal Supplements},
