### 2025-08-04 e455c95
Paper edits.
