+++ paper.md
@@ -26,7 +26,7 @@ SWIFTGalaxy is an open-source astrophysics module that extends SWIFTSimIO [@Borr
 
 # Background
 
-Cosmological hydrodynamical galaxy formation simulations begin with a representation of the early universe where its two key constituents - dark matter and gas - are discretized. The dark matter is usually represented with collisionless particles that respond only to the gravitational force, while gas may be represented by particles (Lagrangian tracers) or mesh cells (Eulerian tracers) and obeys the laws of hydrodynamics in addition to gravity. The SWIFT [@Schaller2024] code takes the former approach, implementing the smoothed particle hydrodynamics [SPH, @Lucy1977; @Gingold1977] formalism. Observations of the cosmic microwave background are used to constrain a multiscale Gaussian random field to formulate the initial conditions for a simulation of a representative region of the early Universe, before the first galaxies formed [e.g. @Bertschinger2001]. In addition to hydrodynamics and gravity, galaxy formation models include additional physical processes such as radiative gas cooling, star formation, energetic feedback from supernovae and supermassive black hole accretion, and more. Many of these are formally unresolved, leading these to be collectively refered to as "sub-grid" elements of the models.
+Cosmological hydrodynamical galaxy formation simulations begin with a representation of the early universe where its two key constituents - dark matter and gas - are discretized. The dark matter is usually represented with collisionless particles that respond only to the gravitational force, while gas may be represented by particles (Lagrangian tracers) or mesh cells (Eulerian tracers) and obeys the laws of hydrodynamics in addition to gravity. The SWIFT [@Schaller2024] code takes the former approach, implementing the smoothed particle hydrodynamics [SPH; @Lucy1977; @Gingold1977] formalism. Observations of the cosmic microwave background are used to constrain a multiscale Gaussian random field to formulate the initial conditions for a simulation of a representative region of the early Universe, before the first galaxies formed [e.g. @Bertschinger2001]. In addition to hydrodynamics and gravity, galaxy formation models include additional physical processes such as radiative gas cooling, star formation, energetic feedback from supernovae and supermassive black hole accretion, and more. Many of these are formally unresolved, leading these to be collectively refered to as "sub-grid" elements of the models.
 
 As time integration of a cosmological hydrodynamical galaxy formation simulation proceeds, outputs containing a "snapshot" of the simulation state are typically written at regular intervals. These contain tables of the properties of particles including at minimum positions, velocities, masses and unique identifiers. More physically complex particle types, such as stars and gas, carry additional properties such as temperature, metallicity, internal energy, and many more. SWIFT snapshot files also include rich metadata recording the full configuration of the code and galaxy formation model, plus physical units and related information for all data.
 
@@ -42,7 +42,7 @@ A core design principle of SWIFTGalaxy is that only operations with an unambiguo
 
 Despite being tailored for use with SWIFT simulations, SWIFTGalaxy is modular: it has interchangeable interfaces to different halo catalogue formats. In addition to the preferred SOAP format, the Velociraptor [@Elahi2019] and [Caesar](https://github.com/dnarayanan/caesar) formats are also supported, and adding support for other formats is straightforward by design. A "standalone" mode is also available so that the other features of SWIFTGalaxy can be used on arbitrary collections of particles even if no halo catalogue is available or desired.
 
[…]
