### 2025-08-04 ef1ebb7
Remove test workflows for paper branch.
