+++ .github/workflows/python-publish.yml
@@ -1,39 +1,0 @@
-# This workflow will upload a Python Package using Twine when a release is created, or a manual trigger
-# For more information see: https://help.github.com/en/actions/language-and-framework-guides/using-python-with-github-actions#publishing-to-package-registries
-
-name: Upload Python Package
-
-on:
-  release:
-    types: [published]
-  workflow_dispatch:      
-
-jobs:
-  lint-and-test:
-    uses: ./.github/workflows/lint_and_test.yml
-  pypi-publish:
-    name: Upload release to PyPI
-    needs: [lint-and-test]
-    runs-on: ubuntu-latest
-    environment:
-      name: pypi
-      url: https://pypi.org/p/swiftgalaxy
-    permissions:
-      id-token: write
-    steps:
-    - uses: actions/checkout@v4
-    - name: Set up Python
-      uses: actions/setup-python@v5
-      with:
-        python-version: '3.12'
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install setuptools wheel build
-        pip install -r requirements.txt
-        pip install .
-    - name: Build
-      run: |
-        pyproject-build
-    - name: Publish package distributions to PyPI
-      uses: pypa/gh-action-pypi-publish@release/v1
