+++ paper.md
@@ -22,11 +22,11 @@ bibliography: bibliography.bib
 
 # Summary
 
-SWIFTGalaxy is an open-source astrophysics module that extends SWIFTSimIO [@Borrow2020] tailored to analyses of particles belonging to individual galaxies simulated with SWIFT [@Schaller2024]. It inherits from and extends the functionality of the SWIFTSimIO's SWIFTDataset class. It understands the content of halo catalogues and therefore which particles belong to a galaxy or other group of particles, and its integrated properties. The particles occupy a coordinate frame that is enforced to be consistent, such that particles loaded on-the-fly will match e.g. rotations and translations of particles already in memory. Intuitive masking of particle datasets is also enabled. Utilities to make working in cylindrical and spherical coordinate systems more convenient are also provided. Finally, tools to iterate efficiently over multiple galaxies are provided.
+SWIFTGalaxy is an open-source astrophysics module that extends SWIFTSimIO [@Borrow2020] tailored to analyses of particles belonging to individual galaxies simulated with SWIFT [@Schaller2024]. It inherits from and extends the functionality of SWIFTSimIO's SWIFTDataset class. It understands the content of halo catalogues and therefore which particles belong to a galaxy or other group of particles, and its integrated properties. The particles occupy a coordinate frame that is enforced to be consistent, such that particles loaded on-the-fly will match e.g. rotations and translations of particles already in memory. Intuitive masking of particle datasets is also enabled. Utilities to make working in cylindrical and spherical coordinate systems more convenient are also provided. Finally, tools to iterate efficiently over multiple galaxies are provided.
 
 # Background
 
-Cosmological hydrodynamical galaxy formation simulations begin with a representation of the early universe where its two key constituents - dark matter and gas - are discretized. The dark matter is usually represented with collisionless particles that respond only to the gravitational force, while gas may be represented by particles (Lagrangian tracers) or mesh cells (Eulerian tracers) and obeys the laws of hydrodynamics in addition to gravity. The SWIFT [@Schaller2024] code takes the former approach, implementing the smoothed particle hydrodynamics [SPH; @Lucy1977; @Gingold1977] formalism. Observations of the cosmic microwave background are used to constrain a multiscale Gaussian random field to formulate the initial conditions for a simulation of a representative region of the early Universe, before the first galaxies formed [e.g. @Bertschinger2001]. In addition to hydrodynamics and gravity, galaxy formation models include additional physical processes such as radiative gas cooling, star formation, energetic feedback from supernovae and supermassive black hole accretion, and more. Many of these are formally unresolved, leading these to be collectively refered to as "sub-grid" elements of the models.
+Cosmological hydrodynamical galaxy formation simulations begin with a representation of the early universe where its two key constituents - dark matter and gas - are discretized. The dark matter is usually represented with collisionless particles that respond only to the gravitational force, while gas may be represented by particles (Lagrangian tracers) or mesh cells (Eulerian tracers) and obeys the laws of hydrodynamics in addition to gravity. The SWIFT [@Schaller2024] code takes the former approach, implementing the smoothed particle hydrodynamics [SPH; @Lucy1977; @Gingold1977] formalism. Observations of the cosmic microwave background are used to constrain a multiscale Gaussian random field to formulate the initial conditions for a simulation of a representative region of the early Universe, before the first galaxies formed [e.g. @Bertschinger2001]. In addition to hydrodynamics and gravity, galaxy formation models include additional physical processes such as radiative gas cooling, star formation, energetic feedback from supernovae and supermassive black hole accretion, and more. Many of these are formally unresolved, leading these to be collectively referred to as "sub-grid" elements of the models.
 
[…]
