### 2025-08-04 505bdb7
Paper edits.
