+++ paper.md
@@ -28,9 +28,9 @@ SWIFTGalaxy is an open-source astrophysics module that extends SWIFTSimIO [@Borr
 
 Cosmological hydrodynamical galaxy formation simulations begin with a representation of the early universe where its two key constituents - dark matter and gas - are discretized. The dark matter is usually represented with collisionless particles that respond only to the gravitational force, while gas may be represented by particles (Lagrangian tracers) or mesh cells (Eulerian tracers) and obeys the laws of hydrodynamics in addition to gravity. The SWIFT [@Schaller2024] code takes the former approach, implementing the smoothed particle hydrodynamics [SPH; @Lucy1977; @Gingold1977] formalism. Observations of the cosmic microwave background are used to constrain a multiscale Gaussian random field to formulate the initial conditions for a simulation of a representative region of the early Universe, before the first galaxies formed [e.g. @Bertschinger2001]. In addition to hydrodynamics and gravity, galaxy formation models include additional physical processes such as radiative gas cooling, star formation, energetic feedback from supernovae and supermassive black hole accretion, and more. Many of these are formally unresolved, leading these to be collectively referred to as "sub-grid" elements of the models.
 
-As time integration of a cosmological hydrodynamical galaxy formation simulation proceeds, outputs containing a "snapshot" of the simulation state are typically written at regular intervals. These contain tables of the properties of particles including at minimum positions, velocities, masses and unique identifiers. More physically complex particle types, such as stars and gas, carry additional properties such as temperature, metallicity, internal energy, and many more. SWIFT snapshot files also include rich metadata recording the full configuration of the code and galaxy formation model, plus physical units and related information for all data.
+As time integration of a cosmological hydrodynamical galaxy formation simulation proceeds, outputs containing a "snapshot" of the simulation state are typically written at regular intervals. These contain tables of the properties of particles including at minimum positions, velocities, masses and unique identifiers. More physically complex particle types, such as stars and gas, carry additional properties such as temperature, metallicity, internal energy, and many more. SWIFT snapshot files include rich metadata recording the full configuration of the code and galaxy formation model, plus physical units and related information for all data.
 
-For many types of analysis the snapshot files alone are insufficient: knowledge of which groups of particles "belong" to gravitationally bound structures - galaxies - is needed. In SWIFT this is most commonly determined in a two-stage process. In the first stage particles are linked to neigbouring particles separated by less than a suitably chosen linking length to form connected "friends-of-friends" [FOF; @Davis1985] groups. The second step consists of finding overdense, self-bound particle groups or "halos". There are numerous algorithms that accomplish this task; the current implementation recommended for the SWIFT community is the HBT-Herons [@ForouharMoreno2025] implementation of the hierarchical bound-tracing plus [HBT+; @Han2018]. Finally, the properties of halos can be tabulated in "halo catalogues" - again many algorithms exist. This is often done as part of the halo finding process, but the recommendation for the SWIFT community is to move this to a separate step and use the spherical overdensity aperture properties [SOAP; @McGibbon2025] tool.
[…]
