### 2025-08-04 SWIFTSIM/swiftgalaxy#56 @kyleaoman
Thanks for flagging it, it was missing by design (just because it hadn't made it to the top of my to-do list yet and isn't strictly needed until the package is sent to JOSS). There is now a draft JOSS paper in the root directory of the [`joss-paper`](https://github.com/SWIFTSIM/swiftgalaxy/tree/joss-paper) branch.
