### 2025-08-04 SWIFTSIM/swiftgalaxy#53 @kyleaoman
Apparently solved in https://github.com/dnarayanan/caesar/pull/104, would be good to have confirmation from @blakeaw
