### 2025-08-04 #244 @kyleaoman
@blakeaw I think that I've now addressed all of the issues that you've raised so far.

SWIFTSIM/swiftgalaxy#53 is still open but it seems like the caesar maintainers have picked it up and fixed it (https://github.com/dnarayanan/caesar/pull/104). Would be good to know if this has actually fixed the issue (for me caesar installed cleanly both before and after their fix...), before I close the issue.

SWIFTSIM/swiftgalaxy#55 was primarily a symptom of how I was handling temporary file creation/destruction (i.e. very badly). That's now much improved, but would be good to know if the test suite now passes on your system or if there are any remaining issues, so have also left that open for the time being.

Not sure if your review was otherwise complete or if there may be more coming?

@crhea93 I think that the second review is now nearly a month overdue? It's the time of year when people leave on holiday so I'm not trying to put a big rush on things, but don't want this to sit in limbo for ever, either.
