### 2025-09-12 #244 @blakeaw
> [@blakeaw](https://github.com/blakeaw) I think that I've now addressed all of the issues that you've raised so far.
>
> [SWIFTSIM/swiftgalaxy#53](https://github.com/SWIFTSIM/swiftgalaxy/issues/53) is still open but it seems like the caesar maintainers have picked it up and fixed it ([dnarayanan/caesar#104](https://github.com/dnarayanan/caesar/pull/104)). Would be good to know if this has actually fixed the issue (for me caesar installed cleanly both before and after their fix...), before I close the issue.
>
> [SWIFTSIM/swiftgalaxy#55](https://github.com/SWIFTSIM/swiftgalaxy/issues/55) was primarily a symptom of how I was handling temporary file creation/destruction (i.e. very badly). That's now much improved, but would be good to know if the test suite now passes on your system or if there are any remaining issues, so have also left that open for the time being.
>
> Not sure if your review was otherwise complete or if there may be more coming?

Hi @kyleaoman, thanks for addressing the various issues.

I've confirmed that Caesar installs cleanly now for me as well, so I went ahead and closed that issue.

Unfortunately, I'm still having issues on my Windows system with the temporary file handling when running the test suite - commented in the appropriate issue.

Otherwise, I think I just need to look over the JOSS paper to complete my review.  I'll try to take care of that in the next few days.
