### 2025-09-17 SWIFTSIM/swiftgalaxy#64 @blakeaw
pyOpenSci review - JOSS paper in prep references
This issue is part of this package's pyOpenSci review at https://github.com/pyOpenSci/software-submission/issues/244

**Severity:** Minor

## Description

Regarding checklist item:
> References: With DOIs for all those that have one (e.g. papers, datasets, software).

I think the two _in preparation_ references
> especially in the context of the COLIBRE simulations project (Schaye et al., in preparation; Chaikin et al., in preparation)

can be included in `bibliography.bib` as `@unpublished` entries and referenced accordingly in the text so that they are included in the complete reference list when the PDF is compiled for JOSS.
