### 2025-09-17 SWIFTSIM/swiftgalaxy#64 @kyleaoman
Thanks - I knew these were imminent when I drafted the text and they're on arxiv now, so I'll update the references.
