### 2025-09-23 #244 @cmillion
@kyleaoman Would you consider a modification to the data retrieval method in `demo_data.py` to use something more cross-platform like `requests` instead of `wget`? The issue being that the demos don't run on systems that don't have `wget`, like OSX. If I hack around it, then the code otherwise works on OSX (afaict). Alternatively, specify in the Quickstart or elsewhere that the demos won't work on systems that don't have `wget`... but it seems like running through the demos is really important for understanding what this library does and how, so that one hiccup just seems to create an unnecessary barrier to use.
