### 2025-09-23 #244 @kyleaoman
> Would you consider a modification to the data retrieval method in `demo_data.py` to use something more cross-platform like `requests` instead of `wget`?

This is a good suggestion - could you open an issue?
