### 2025-09-23 SWIFTSIM/swiftgalaxy#65 @cmillion
Modify demo data retrieval to use something other than wget
The issue being that the demos don't run on systems that don't have wget, like OSX. If I hack around it, then the code otherwise works on OSX (afaict). This only affects the demos. It seems like running through the demos is really important for understanding what this library does and how, so that one hiccup just seems to create an unnecessary barrier to use.

The specific location of the code is the `wget` call in demo_data.py

I suggest using `requests` to grab the data.

I can make a PR for this if the contribution is desired.

Cross reference to [the pyOpenSci review thread for this package.](https://github.com/pyOpenSci/software-submission/issues/244#issuecomment-3322038603)
