### 2025-09-26 e4a2ddb
Replace wget call with requests.
