+++ bibliography.bib
@@ -207,3 +207,38 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2001ApJS..137....1B},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{Schaye2025,
+       author = {{Schaye}, Joop and {Chaikin}, Evgenii and {Schaller}, Matthieu and {Ploeckinger}, Sylvia and {Hu{\v{s}}ko}, Filip and {McGibbon}, Rob and {Trayford}, James W. and {Ben{\'\i}tez-Llambay}, Alejandro and {Correa}, Camila and {Frenk}, Carlos S. and {Richings}, Alexander J. and {Forouhar Moreno}, Victor J. and {Bah{\'e}}, Yannick M. and {Borrow}, Josh and {Durrant}, Anna and {Gebek}, Andrea and {Helly}, John C. and {Jenkins}, Adrian and {Lacey}, Cedric G. and {Ludlow}, Aaron and {Nobels}, Folkert S.~J.},
+        title = "{The COLIBRE project: cosmological hydrodynamical simulations of galaxy formation and evolution}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics of Galaxies, Cosmology and Nongalactic Astrophysics},
+         year = 2025,
+        month = aug,
+          eid = {arXiv:2508.21126},
+        pages = {arXiv:2508.21126},
+          doi = {10.48550/arXiv.2508.21126},
+archivePrefix = {arXiv},
+       eprint = {2508.21126},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2025arXiv250821126S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Chaikin2025,
+       author = {{Chaikin}, Evgenii and {Schaye}, Joop and {Schaller}, Matthieu and {Ploeckinger}, Sylvia and {Bah{\'e}}, Yannick M. and {Ben{\'\i}tez-Llambay}, Alejandro and {Correa}, Camila and {Forouhar Moreno}, Victor J. and {Frenk}, Carlos S. and {Hu{\v{s}}ko}, Filip and {Kugel}, Roi and {McGibbon}, Robert and {Richings}, Alexander J. and {Trayford}, James W. and {Borrow}, Josh and {Crain}, Robert A. and {Helly}, John C. and {Lacey}, Cedric G. and {Ludlow}, Aaron and {Nobels}, Folkert S.~J.},
+        title = "{COLIBRE: calibrating subgrid feedback in cosmological simulations that include a cold gas phase}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics of Galaxies},
+         year = 2025,
+        month = sep,
+          eid = {arXiv:2509.04067},
+        pages = {arXiv:2509.04067},
+          doi = {10.48550/arXiv.2509.04067},
+archivePrefix = {arXiv},
+       eprint = {2509.04067},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2025arXiv250904067C},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
