### 2025-09-26 075e914
Replace wget call with requests (#66)
