+++ .github/workflows/lint_and_test.yml
@@ -108,3 +108,61 @@ jobs:
       uses: codecov/codecov-action@v5
       env:
         CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
+
+  windows-tests:
+    needs: lint
+    runs-on: windows-latest
+    strategy:
+      matrix:
+        python-version: ["3.12"]
+
+    steps:
+    - name: Checkout swiftgalaxy
+      uses: actions/checkout@v4
+    - name: Setup MPI
+      uses: mpi4py/setup-mpi@v1 
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v5
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Cache pip
+      uses: actions/cache@v4
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install flake8 pytest pytest-xdist
+        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+        if [ -f optional_requirements.txt ]; then pip install -r optional_requirements.txt; fi
+        pip install -e .
+    - name: Test with pytest
+      run: |
+        pytest -n 2
+
+  macos-tests:
+    needs: lint
+    runs-on: macos-latest
+    strategy:
+      matrix:
+        python-version: ["3.12"]
+
+    steps:
+    - name: Checkout swiftgalaxy
+      uses: actions/checkout@v4
+    - name: Setup MPI
+      uses: mpi4py/setup-mpi@v1 
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v5
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Cache pip
+      uses: actions/cache@v4
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install flake8 pytest pytest-xdist
+        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+        if [ -f optional_requirements.txt ]; then pip install -r optional_requirements.txt; fi
+        pip install -e .
+    - name: Test with pytest
+      run: |
+        pytest -n 2
