+++ .github/workflows/lint_and_test.yml
@@ -129,8 +129,8 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install flake8 pytest pytest-xdist
-        if exist requirements.txt pip install -r requirements.txt
-        if exist optional_requirements.txt pip install -r optional_requirements.txt
+        if ( exist requirements.txt ) pip install -r requirements.txt
+        if ( exist optional_requirements.txt ) pip install -r optional_requirements.txt
         pip install -e .
     - name: Test with pytest
       run: |
