+++ .github/workflows/lint_and_test.yml
@@ -129,8 +129,8 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install flake8 pytest pytest-xdist
-        if ( [System.IO.File]::Exists(requirements.txt) ) { pip install -r requirements.txt }
-        if ( [System.IO.File]::Exists(optional_requirements.txt) ) { pip install -r optional_requirements.txt }
+        if ( Test-Path requirements.txt -PathType Leaf ) { pip install -r requirements.txt }
+        if ( Test-Path optional_requirements.txt -PathType Leaf ) { pip install -r optional_requirements.txt }
         pip install -e .
     - name: Test with pytest
       run: |
