### 2025-09-26 7ef368c
Get rid of unused MPI in CI.
