+++ .github/workflows/lint_and_test.yml
@@ -47,8 +47,6 @@ jobs:
     steps:
     - name: Checkout swiftgalaxy
       uses: actions/checkout@v4
-    - name: Setup MPI
-      uses: mpi4py/setup-mpi@v1 
     - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v5
       with:
@@ -79,8 +77,6 @@ jobs:
     steps:
     - name: Checkout swiftgalaxy
       uses: actions/checkout@v4
-    - name: Setup MPI
-      uses: mpi4py/setup-mpi@v1 
     - name: Set up Python 3.12
       uses: actions/setup-python@v5
       with:
@@ -119,8 +115,6 @@ jobs:
     steps:
     - name: Checkout swiftgalaxy
       uses: actions/checkout@v4
-    - name: Setup MPI
-      uses: mpi4py/setup-mpi@v1 
     - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v5
       with:
@@ -146,8 +140,6 @@ jobs:
     steps:
     - name: Checkout swiftgalaxy
       uses: actions/checkout@v4
-    - name: Setup MPI
-      uses: mpi4py/setup-mpi@v1 
     - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v5
       with:
