### 2025-09-26 SWIFTSIM/swiftgalaxy#55 @kyleaoman
I switched on running the test suite on a windows platform on github and indeed I reproduce the test failures. I do have a windows desktop at home, don't usually use it for any development but will see if I can run the suite there and troubleshoot...
