### 2025-09-29 SWIFTSIM/swiftgalaxy#55 @kyleaoman
It turns out this is an upstream issue:

https://github.com/SWIFTSIM/swiftsimio/issues/257
