### 2025-10-09 #244 @kyleaoman
The last of the issues raised so far is now [basically solved](https://github.com/SWIFTSIM/swiftgalaxy/pull/67) (pending some code reviews, merges and publishing releases). @cmillion @blakeaw it would be appreciated if we could bring the review to a conclusion in the next couple of weeks, if possible.
