### 2025-10-16 2242adc
Update index.rst
