### 2025-10-16 #244 @cmillion
The notebook `examples/SWIFTGalaxy_Colibre_QuickStart.ipynb` appears to be pointing toward simulation files that are supposed to be obtained or produced externally to the notebook file itself. The way in which one goes about producing those files is not clear to me. Could you please help me understand how to be able to run this example?

I _think_ that I would need to run COLIBRE / cosma in order to generate example files in order to run this notebook. Per the statement "This guide is for users working with Colibre simulation outputs on the cosma system." If that is the case, could small example files be included or made available for download? Or can the notebook be updated to clearly indicated that it _cannot_ be run without such files.

---

The document states

> HBT-Herons records which particles belong to each individual halo in a set of "membership" files, usually found alongside the halo catalogue in a subdirectory, in this case `membership_0127/membership_0127.X.hdf5` (where X is replaced by integers, each corresponding to one part of the "raw" snapshot). `swiftgalaxy` expects to find the information contained in these files directly in the (single, monolithic) simulation snapshot file. Such a snapshot file exists alongside the SOAP catalogue as `colibre_with_SOAP_membership_0127.hdf5`.

and then assembles a path with

```
colibre_base_path = "/cosma8/data/dp004/colibre/Runs/"
simulation_dir = "L0025N0752/THERMAL_AGN_m5"
soap_catalogue_file = os.path.join(
    colibre_base_path,
    simulation_dir,
    "SOAP/halo_properties_0127.hdf5",
)

virtual_snapshot_file = os.path.join(
    colibre_base_path, simulation_dir, "SOAP/colibre_with_SOAP_membership_0127.hdf5"
)
```
