### 2025-10-16 #244 @cmillion
My final review is below. The issues I found are, IMO, minor documentation issues at worst. -CM

---

## Package Review

- ✅ As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

#### Documentation

- ✅ **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- ✅ **Installation instructions:** for the development version of the package and any non-standard dependencies in README.
- ✅ **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
- ✅ **Function Documentation:** for all user-facing functions.
- ✅ **Examples** for all user-facing functions.
- ✅ **Community guidelines** including contribution guidelines in the README or CONTRIBUTING.
- ✅ **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

Readme file  requirements

- ✅ Package has a README.md file in the root directory.

**Reviewer note: It is a README.rst**

- ✅ The package name
    - ✅ Continuous integration and test coverage,
    - ✅ Docs building (if you have a documentation website),
    - ✅ Python versions supported,
    - ✅ Current package version (on PyPI / Conda).

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

- ✅ Short description of package goals.
- ✅ Package installation instructions
- ✅ Any additional setup required to use the package (authentication tokens, etc.)
- ✅ Descriptive links to all vignettes. If the package is small, there may only be a need for one vignette which could be placed in the README.md file.
    - ✅ Brief demonstration of package usage (as it makes sense - links to vignettes could also suffice here if package description is clear)
- ✅ Link to your documentation website.
- ✅ If applicable, how the package compares to other similar packages and/or how it relates to other packages in the scientific ecosystem.
- ✅ Citation information

#### Usability

The package structure should follow the general community best practices. In general, please consider whether:

- ✅ Package documentation is clear and easy to find and use.
- ✅ The need for the package is clear
- ✅ All functions have documentation and associated examples for use
- ✅ The package is easy to install

#### Functionality

- ✅ **Installation:** Installation succeeds as documented.
- ✅ **Functionality:** Any functional claims of the software been confirmed.
- ✅ **Performance:** Any performance claims of the software been confirmed.
- ✅ **Automated tests:**
  - ✅ All tests pass on the reviewer's local machine for the package version submitted by the author. Ideally this should be a tagged version making it easy for reviewers to install.
  - ✅ Tests cover essential functions of the package and a reasonable range of inputs and conditions.
- ✅ **Continuous Integration:** Has continuous integration setup (We suggest using Github actions but any CI platform is acceptable for review)
- ✅ **Packaging guidelines**: The package conforms to the pyOpenSci [packaging guidelines](https://www.pyopensci.org/python-package-guide).
    A few notable highlights to look at:
    - ✅ Package supports modern versions of Python and not [End of life versions](https://endoflife.date/python).
    - ✅ Code format is standard throughout package and follows PEP 8 guidelines (CI tests for linting pass)

- ✅ The package has an **obvious research application** according to JOSS's definition in their [submission requirements](http://joss.theoj.org/about#submission_requirements).

- ✅ **A short summary** describing the high-level functionality of the software
- ✅ **Authors:** A list of authors with their affiliations
- ✅ **A statement of need** clearly stating problems the software is designed to solve and its target audience.
- ✅ **References:** With DOIs for all those that have one (e.g. papers, datasets, software).

#### Final approval (post-review)

- ✅ **The author has responded to my review and made changes to my satisfaction. I recommend approving this package.**

Estimated hours spent reviewing: 3.5

---

#### Review Comments

* The software appears to not support the (very new) Python 3.14, and the documentation and badge should be updated to reflect that.
* The `examples/SWIFTGalaxy_Colibre_QuickStart.ipynb` example requires external data files to run which are not provided, and it is not entirely clear how to obtain them. Further documentation of this would probably be a good idea.
* In order for all of the tests to pass, it appears to be the case that every _optional_ dependency must be installed. This is fine (and makes sense, because functionality requiring those dependencies should also have test coverage), but maybe a note in the dependency section of the docs would be helpful to clarity.
