### 2025-10-17 #244 @kyleaoman
@cmillion thanks for the review, I'll fix those issues. Indeed the colibre notebook requires files that I'm not allowed to distribute, but many people in that collaboration use the package and I wanted to give an example since there are some particularities. I'll just clarify in the preamble of the notebook. Those data will become public later, I will of course update the notebook with file access instructions when that happens.

Could I just check: the submission includes a JOSS paper, did you look at that? The draft can currently be found under artifacts at: https://github.com/SWIFTSIM/swiftgalaxy/actions/runs/18033400592
