### 2025-10-17 #244 @crhea93
@kyleaoman Fantastic! I think we're nearly done then :)

As for the JOSS paper, once your package has been accepted by PyOpenSci (i.e. these reviews are complete), we will ask you to submit your package to JOSS where it will undergo an expedited review assuming it falls under their scope requirements (which I think it certainly will). At that point, the JOSS folks will review your paper :)

I'll give more in depth details once our reviews are completed! More details can be found here https://www.pyopensci.org/software-peer-review/our-process/review-timeline.html#joss-submission
