### 2025-10-17 #244 @cmillion
@kyleaoman I missed the presence of a JOSS paper. Thanks. I have reviewed it and updated my review checklist inplace above.
