### 2025-10-17 #244 @kyleaoman
@cmillion on specifying python<3.14 in the package metadata, I think it's best not to do this, but it's a bit of an unsolved problem, see e.g. https://discuss.python.org/t/requires-python-upper-limits/12663

I *could* look at what currently breaks in 3.14 but I fully expect this to be things in `unyt`, `swiftsimio`, possibly other dependencies (`numba`!) and, possibly, some things also in `swiftgalaxy`. It will be supported in due course but I hope we can agree this is out of scope for this review, and I'd rather not set an explicit upper version limit.
