### 2025-10-17 #244 @kyleaoman
@cmillion on optional dependencies, I think it's noted clearly in the right section of the docs (https://swiftgalaxy.readthedocs.io/en/latest/getting_started/index.html#installation-for-development). I've now linked to this from https://github.com/SWIFTSIM/swiftgalaxy?tab=contributing-ov-file, which is, I suspect, where you found it lacking?
