### 2025-10-17 #244 @kyleaoman
> I don't quite understand. Are you able to get the software to run on 3.14 without modification? I was not. Setting an explicit upper limit version now does not prevent you from expanding it in the future when it is supported.
>
> However, I won't stake out a strong position on this. Whatever you choose, this package passes my review. My comments were all minor notes.

I agree that it doesn't work on 3.14, but whether you should set an upper limit in `requires-python` is a bit of a controversial question - you'll get different answers from different (knowledgeable) people. For now I prefer not to, I think.
