### 2025-10-17 #244 @kyleaoman
@cmillion I've addressed your points with a couple of commits to main, if you're satisfied then please tick the last box and @crhea93 can pick this up. Thanks again for reviewing!
