### 2025-10-17 #244 @cmillion
I had to go educate myself on this controversy. I don't think there's a consensus. Whatever you choose re: version pinning is fine with me.

I have checked the final box on my review.
