### 2025-10-17 #244 @crhea93
---

🎉 `swiftgalaxy` has been approved by pyOpenSci! Thank you @kyleaoman  for submitting `swiftgalaxy` and many thanks to @cmillion and @blakeaw  for reviewing this package! 😸

- [ ] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

- [ ] <Editor Task> Once the JOSS issue is opened for the package, we strongly suggest that you subscribe to issue updates. This will allow you to continue to update the issue labels on this review as it goes through the JOSS process.

🎉 Congratulations! You are now published with both JOSS and pyOpenSci! 🎉

- [x] Change the status tag of the issue to `6/pyOS-approved6 🚀🚀🚀`.  (You can remove all other labels; If the the package is moving on to JOSS then add a JOSS label)
- [x] Update the header of this issue with the date accepted, the version accepted, and the accepted version DOI (from Zenodo) to the header of this issue.
- [x] Invite the package maintainer(s) and both reviewers to the pyOpenSci Slack.
- [x]  If the review is complete (not submitting to JOSS), close the issue when all items on this checklist are complete. If it's moving to JOSS, please close the issue once the JOSS review is complete and you've performed the task below.

---
