### 2025-10-17 #244 @crhea93
@kyleaoman I would like to invite you and your maintainer team to write a blog post (totally optional) on your package for us to promote your work! if you are interested - here are a few examples of other blog posts:

[pandera](https://www.pyopensci.org/blog/pandera-python-pandas-dataframe-validation)
[movingpandas](https://www.pyopensci.org/blog/movingpandas-movement-data-analysis)

and [here](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_posts/2019-11-18-pandera-dataframe-validation.md) is a markdown example that you could use as a guide when creating your post.

This is certainly not mandatory, but we believe it will help increase the visibility of the package :D
