### 2025-10-17 #244 @crhea93
@kyleaoman @cmillion @blakeaw Thank you for all the wonderful work! I wanted to take a  moment to invite you to join our PyOpenSci slack if you are not already a member! If this is of interest to you, please let me know :)
