### 2025-10-17 b6ef3cf
Add pyopensci badge!
