### 2025-10-17 #244 @kyleaoman
@crhea93 just to check, it's now my job to fill out a JOSS review form, right?

Other than that, I've added the badge and will mint a new release soon :)
