### 2025-10-17 42fbdb9
Revert badges to md in readme, PyPI won't render raw html.
