### 2025-10-17 f5006a2
Update README.rst
