### 2025-10-20 214d260
Initial editorial tweaks for JOSS publication
