+++ bibliography.bib
@@ -1,6 +1,6 @@
 @ARTICLE{Borrow2020,
        author = {{Borrow}, Josh and {Borrisov}, Alexei},
-        title = "{swiftsimio: A Python library for reading SWIFT data}",
+        title = {{swiftsimio}: A {Python} library for reading {SWIFT} data},
       journal = {The Journal of Open Source Software},
      keywords = {Python, cosmology, i/o, simulations, astronomy},
          year = 2020,
@@ -16,7 +16,7 @@
 
 @ARTICLE{Schaller2024,
        author = {{Schaller}, Matthieu and {Borrow}, Josh and {Draper}, Peter W. and {Ivkovic}, Mladen and {McAlpine}, Stuart and {Vandenbroucke}, Bert and {Bah{\'e}}, Yannick and {Chaikin}, Evgenii and {Chalk}, Aidan B.~G. and {Chan}, Tsang Keung and {Correa}, Camila and {van Daalen}, Marcel and {Elbers}, Willem and {Gonnet}, Pedro and {Hausammann}, Lo{\"\i}c and {Helly}, John and {Hu{\v{s}}ko}, Filip and {Kegerreis}, Jacob A. and {Nobels}, Folkert S.~J. and {Ploeckinger}, Sylvia and {Revaz}, Yves and {Roper}, William J. and {Ruiz-Bonilla}, Sergio and {Sandnes}, Thomas D. and {Uyttenhove}, Yolan and {Willis}, James S. and {Xiang}, Zhen},
-        title = "{SWIFT: A modern highly-parallel gravity and smoothed particle hydrodynamics solver for astrophysical and cosmological applications}",
+        title = {{SWIFT}: A modern highly-parallel gravity and smoothed particle hydrodynamics solver for astrophysical and cosmological applications},
       journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {software: simulations, methods: numerical, software: public release, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Astrophysics of Galaxies, Computer Science - Distributed, Parallel, and Cluster Computing},
          year = 2024,
@@ -34,7 +34,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{Elahi2019,
        author = {{Elahi}, Pascal J. and {Ca{\~n}as}, Rodrigo and {Poulton}, Rhys J.~J. and {Tobar}, Rodrigo J. and {Willis}, James S. and {Lagos}, Claudia del P. and {Power}, Chris and {Robotham}, Aaron S.~G.},
-        title = "{Hunting for galaxies and halos in simulations with VELOCIraptor}",
+        title = {Hunting for galaxies and halos in simulations with {VELOCIraptor}},
       journal = {Publications of the Astronomical Society of Australia},
      keywords = {dark matter, galaxies: evolution, galaxies: halos, methods: numerical, Astrophysics - Cosmology and Nongalactic Astrophysics},
          year = 2019,
@@ -52,7 +52,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{McGibbon2025,
        author = {{McGibbon}, Robert and {Helly}, John and {Schaye}, Joop and {Schaller}, Matthieu and {Vandenbroucke}, Bert},
-        title = "{SOAP: A Python Package for Calculating the Properties of Galaxies and Halos Formed in Cosmological Simulations}",
+        title = {{SOAP}: A {Python} package for calculating the properties of galaxies and halos formed in cosmological simulations},
       journal = {The Journal of Open Source Software},
      keywords = {astronomy, Python},
          year = 2025,
[…]
