### 2025-10-23 d877030
Add JOSS reference to README.
