### 2025-10-23 ee870cc
Revert "Switch to branch of swiftsimio for testing."
This reverts commit 55eee6284dccab85383d9b188e7a79d4688f72c4.
