+++ requirements.txt
@@ -1,6 +1,5 @@
 numpy>=2.0.0
 scipy
 unyt
-#swiftsimio>=10.0.0
-git+https://github.com/SWIFTSIM/swiftsimio.git@close_handle
+swiftsimio>=10.0.0
 requests
