+++ .github/workflows/lint_and_test.yml
@@ -60,8 +60,6 @@ jobs:
         key: ${{ runner.os }}-pip-${{ hashFiles('requirements.txt') }}
         restore-keys: |
           ${{ runner.os }}-pip-${{ runner.os }}-
-    - name: Setup fortran compiler for dependencies
-      uses: fortran-lang/setup-fortran@v1
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
@@ -121,6 +119,8 @@ jobs:
       uses: actions/setup-python@v5
       with:
         python-version: ${{ matrix.python-version }}
+    - name: Setup fortran compiler for dependencies
+      uses: fortran-lang/setup-fortran@v1
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
