### 2025-11-10 b1ba6ae
Set fortran compiler in env.
